% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_runs.R, R/action_status.R,
%   R/action_workflows.R
\name{action}
\alias{action}
\alias{action_runs}
\alias{action_status}
\alias{action_workflows}
\title{Return a data frame containing details on a repository workflow runs.}
\usage{
action_runs(repo, limit = 30)

action_status(repo, branch = NULL, before = NULL, limit = 50)

action_workflows(repo, full = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{limit}{Numeric. Maximum number of workflow runs to return. If runs are
missing try increasing this limit, if performance is slow try decreasing it.}

\item{branch}{Character. Vector of branches to report.}

\item{before}{Datetime. Filter all runs created after this timestamp.
Character values are converted using \code{lubridate::ymd_hms()}.}

\item{full}{Logical. Should all workflow columns be returned}
}
\description{
Return a data frame containing details on a repository workflow runs.

Return a data frame with the status of the most recent workflow runs for each repo.

Return a data frame containing details on a repository workflows.
}
\examples{
\dontrun{
action_runs("rundel/ghclass")
}

\dontrun{
action_status(c("rundel/ghclass", "rundel/parsermd"))
}

\dontrun{
action_workflows("rundel/ghclass")
}

}
