% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{add_content}
\alias{add_content}
\title{Modify a file within a repository}
\usage{
add_content(repo, file, content, after = NULL,
  message = "Added content", branch = "master")
}
\arguments{
\item{repo}{Character. Address of repository in `owner/name`` format.}

\item{file}{Character. File's path within the repository.}

\item{content}{Character. Content to be added to the file.}

\item{after}{Character. Regex pattern, if not \code{NULL} content will be inserted directly after the first match.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch to use, defaults to "master".}
}
\description{
Modify a file within a repository
}
\seealso{
Other file functions: \code{\link{add_file}},
  \code{\link{get_file}}, \code{\link{get_readme}},
  \code{\link{put_file}}
}
\concept{file functions}
