% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{add_file}
\alias{add_file}
\alias{grab_repos}
\title{Add files to a repo}
\usage{
add_file(repo, message, file, branch = "master", preserve_path = FALSE,
  overwrite = FALSE, include_admin = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{message}{Character. Commit message.}

\item{file}{Character. Local file paths of file or files to be added.}

\item{branch}{Character. Name of branch to use, defaults to "master".}

\item{preserve_path}{Logical. Should the local relative path be preserved.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to FALSE.}

\item{include_admin}{Logical. Should users with admin privileges be included in checking for previous file modifications, defaults to FALSE.}
}
\description{
\code{add_files} uses the GitHub API to add/update files in an existing repo on GitHub.
}
\examples{
\dontrun{
add_files("rundel/ghclass", "Update DESCRIPTION", "./DESCRIPTION")
}

}
\seealso{
Other local repo functions: \code{\link{add_files-deprecated}}
}
\concept{local repo functions}
