% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{add_files-deprecated}
\alias{add_files-deprecated}
\alias{add_files}
\alias{grab_repos}
\title{Add files to a repo}
\usage{
add_files(repo, message, files, branch = "master", preserve_path = FALSE)
}
\arguments{
\item{repo}{repo names in the form of \emph{owner/name}.}

\item{message}{commit message.}

\item{files}{local file paths of files to be added.}

\item{branch}{name of branch to use, defaults to master.}

\item{preserve_path}{should the local relative path be preserved.}
}
\description{
\code{add_files} uses the GitHub api to add/update files in an existing repo on GitHub.
}
\section{\code{add_files}}{

For \code{add_files}, use \code{\link{add_file}}.
}

\examples{
\dontrun{
add_files("rundel/ghclass", "Update DESCRIPTION", "./DESCRIPTION")
}

}
\seealso{
\code{\link{ourPkg-deprecated}}

Other local repo functions: \code{\link{add_file}}
}
\concept{local repo functions}
\keyword{internal}
