% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{add_files}
\alias{add_files}
\alias{grab_repos}
\title{Add files to a repo}
\usage{
add_files(repo, message, files, branch = "master",
  preserve_path = FALSE)
}
\arguments{
\item{repo}{repo names in the form of \emph{owner/name}.}

\item{message}{commit message.}

\item{files}{local file paths of files to be added.}

\item{branch}{name of branch to use, defaults to master.}

\item{preserve_path}{should the local relative path be preserved.}
}
\description{
\code{add_files} uses the GitHub api to add/update files in an existing repo on GitHub.
}
\examples{
\dontrun{
add_files("rundel/ghclass", "Update DESCRIPTION", "./DESCRIPTION")
}

}
\concept{local repo functions}
