% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_user_to_repo}
\alias{add_user_to_repo}
\alias{add_team_to_repo}
\title{Add a user or team to a repository}
\usage{
add_user_to_repo(repo, user, permission = c("push", "pull", "admin"))

add_team_to_repo(repo, team, permission = c("push", "pull", "admin"))
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{user}{Character. One or more GitHub usernames.}

\item{permission}{Character. Permission to be granted to team for repo ("push", "pull", or "admin"), defaults to "push".

\code{pull} results in read privileges, \code{push} in write privileges,
and \code{admin} in Admin privileges for the team in the respective repository.
Note that permissions will overwrite existing access privileges.}

\item{team}{Character. One or more GitHub team names.}
}
\description{
Add a user or team to a repository
}
\examples{
\dontrun{
add_user_to_repo("ghclass-test/hw1-user01", c("user01", "user02"))
}

}
