% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo.R
\name{add_repos}
\alias{add_repos}
\title{Add changes for local repos via git add}
\usage{
add_repos(repo_dirs, files = ".", git = require_git(), options = "",
  verbose = TRUE)
}
\arguments{
\item{repo_dirs}{Vector of repo directories or a single directory containing one or more repos.}

\item{files}{Vector of files to add}

\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your \code{PATH}, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}
}
\description{
\code{add_repos} uses the git binary to update the provided repositories
using git pull.
}
\examples{
\dontrun{
}

}
\seealso{
Other local repo functions: \code{\link{clone_repos}},
  \code{\link{commit_repos}}, \code{\link{push_repos}},
  \code{\link{update_repos}}
}
