% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{add_team_member}
\alias{add_team_member}
\title{Add Members to an Organizaton's Team(s)}
\usage{
add_team_member(org, user, team, create_missing_teams = TRUE)
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{user}{character, one or more usernames to invite}

\item{team}{character, one or more team names}

\item{create_missing_teams}{logical, if a team does not already exist
should it be added to the GitHub organization.}
}
\description{
\code{add_team_member} add members to GitHub Organization Teams.
}
\examples{
\dontrun{
add_team_member("ghclass-test", "rundel", c("hw1-team01","hw1-team02"))
}

}
\seealso{
Other github organization team related functions: \code{\link{create_team}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_team_members}}, \code{\link{get_teams}},
  \code{\link{rename_team}}
}
\concept{github organization team related functions}
