% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_team_to_repo}
\alias{add_team_to_repo}
\title{Add a team to a repository}
\usage{
add_team_to_repo(repo, team, permission = c("push", "pull", "admin"),
  verbose = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{team}{Character or data frame. Vector of team names.}

\item{permission}{Character. Permission to be granted to team for repo ("push", "pull", or "admin"), defaults to "push".}

\item{verbose}{Logical. Display verbose output.}
}
\description{
\code{add_team_to_repo} adds a team to an existing repository. \code{pull} results in read privileges, \code{push} in write privileges, and \code{admin} in Admin privileges for the team in the respective repository. Note that permissions will overwrite existing access privileges.
}
\examples{
\dontrun{
add_team_to_repo("Sta523-Fa17/resources", c("Team1", "Team2"))
}

}
