% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_args.R
\name{arg_is_chr}
\alias{arg_is_chr}
\title{Internal argument validation helper}
\usage{
arg_is_chr(..., allow_null = FALSE, allow_na = FALSE, allow_empty = TRUE)
}
\description{
Internal argument validation helper
}
\keyword{internal}
