% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_args.R
\name{arg_is_chr_scalar}
\alias{arg_is_chr_scalar}
\title{Internal argument validation helper}
\usage{
arg_is_chr_scalar(..., allow_null = FALSE, allow_na = FALSE)
}
\description{
Internal argument validation helper
}
\keyword{internal}
