% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_args.R
\name{arg_is_pos_int}
\alias{arg_is_pos_int}
\title{Internal argument validation helper}
\usage{
arg_is_pos_int(..., allow_null = FALSE)
}
\description{
Internal argument validation helper
}
\keyword{internal}
