% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_create.R
\name{branch_create}
\alias{branch_create}
\alias{create_branch}
\title{Create branch}
\usage{
branch_create(repo, cur_branch = "master", new_branch)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}

\item{cur_branch}{name of existing branch}

\item{new_branch}{name of branch to create}
}
\description{
\code{branch_create} creates a new branch from an existing git repo.
}
