% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_remove.R
\name{branch_remove}
\alias{branch_remove}
\title{Remove branch}
\usage{
branch_remove(repo, branch)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}

\item{branch}{name of existing branch}
}
\description{
\code{branch_remove} deletes a branch from an existing GitHub repository.
}
