% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_unprotect.R
\name{branch_unprotect}
\alias{branch_unprotect}
\title{Unprotect branch}
\usage{
branch_unprotect(repo, branch = "master")
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}

\item{branch}{name of the branch to unprotect}
}
\description{
\code{branch_unprotect} removes protections from the specified branch. See
\url{https://help.github.com/en/articles/about-protected-branches} for more details
on what this changes.
}
