% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{check_repos-deprecated}
\alias{check_repos-deprecated}
\alias{check_repos}
\title{Check if repository exists}
\usage{
check_repos(repos)
}
\arguments{
\item{repos}{Character. Address of repository in "owner/name" format.}
}
\description{
Check if repository exists
}
\section{\code{check_repos}}{

For \code{check_repos}, use \code{\link{check_repo}}.
}

\seealso{
\code{\link{ourPkg-deprecated}}
}
\keyword{internal}
