% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{check_user_exists}
\alias{check_user_exists}
\title{Check if username(s) exists}
\usage{
check_user_exists(user)
}
\arguments{
\item{user}{Character. Username to be checked. Can be a vector or list of usernames.}
}
\value{
TRUE or FALSE
}
\description{
\code{check_user_exists} returns TRUE if the supplied username(s) exists on GitHub and FALSE otherwise.
}
\examples{
\dontrun{
check_user_exists(c("rundel","hopefullydoesnotexist"))
}

}
