% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo.R
\name{clone_repos}
\alias{clone_repos}
\alias{grab_repos}
\title{Clone repos from github}
\usage{
clone_repos(repos, local_path = "./", branch = "master",
  git = require_git(), options = "", absolute_path = TRUE,
  verbose = FALSE)
}
\arguments{
\item{repos}{Vector of repo names with the form \emph{owner/name}.}

\item{local_path}{Local directory into which the repos will be cloned.}

\item{branch}{Name of branch(es) to clone, defaults to master.}

\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your \code{PATH}, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--bare})}

\item{absolute_path}{Should absolute path be returned for cloned repos.}

\item{verbose}{Display verbose output.}
}
\description{
\code{clone_repos} uses the git binary to clone the provided repositories
into a local directory. The function returns a character vector of paths for the cloned repos.
}
\examples{
\dontrun{
clone_repos("rundel/ghclass")
}

}
\seealso{
Other local repo functions: \code{\link{add_repos}},
  \code{\link{commit_repos}}, \code{\link{push_repos}},
  \code{\link{update_repos}}
}
