% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{create_branch}
\alias{create_branch}
\title{Create branch}
\usage{
create_branch(repo, cur_branch = "master", new_branch)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}

\item{cur_branch}{name of existing branch}

\item{new_branch}{name of branch to create}
}
\description{
\code{create_branch} creates a new branch from an existing git repo.
}
\seealso{
Other branch functions: \code{\link{protect_branch}},
  \code{\link{unprotect_branch}}
}
\concept{branch functions}
