% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{create_file_commit_url}
\alias{create_file_commit_url}
\title{Create file URL to pass to GitHub Commit API}
\usage{
create_file_commit_url(repo, file)
}
\description{
\code{create_file_commit_url} creates a file URL for a single file that can be passed to a \code{gh(METHOD URL)} query. The query is limited to 100 entries.
}
