% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{create_individual_repo}
\alias{create_individual_repo}
\title{Create individual repositories}
\usage{
create_individual_repo(org, user, prefix = "", suffix = "",
  private = TRUE, auto_init = FALSE, gitignore_template = "R")
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{user}{Character or data frame. Listing one or more users.}

\item{prefix}{Character. Resulting repository name will start with this character string.}

\item{suffix}{Character. Resulting repository name will end with this character string.}

\item{private}{Logical. Create private repositories.}

\item{auto_init}{Logical. Initialize the repository with a README.md.}

\item{gitignore_template}{Character. .gitignore template language.}
}
\description{
\code{create_individual_repo} creates repositories for each student for a given
assignment.
}
\examples{
\dontrun{
create_individual_repo("Sta523-Fa17", c("user01","user02"), prefix = "hw01-")
}

}
\seealso{
Other github organization related functions: \code{\link{create_team_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_members-deprecated}},
  \code{\link{get_pending_member}},
  \code{\link{get_repos-deprecated}},
  \code{\link{get_repo}}, \code{\link{get_team_repos}},
  \code{\link{invite_user}}
}
\concept{github organization related functions}
