% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{create_pull_request}
\alias{create_pull_request}
\title{Create pull request}
\usage{
create_pull_request(repo, title, base, head = "master", body = "")
}
\arguments{
\item{repo}{Character. Address of one or more repositories in "owner/name" format.}

\item{title}{Character. Title of the pull request.}

\item{base}{Character. The name of the branch where your changes are implemented. In creating a pull request from a
fork then use \code{username:branch} as the format.}

\item{head}{Character. The branch you want the changed pulled into.}

\item{body}{Character. The text contents of the pull request.}
}
\description{
\code{create_pull_request} creates a pull request on GitHub from the \code{base} branch to the \code{head} branch.
}
