% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{create_repo}
\alias{create_repo}
\title{Create repository}
\usage{
create_repo(org, name, prefix = "", suffix = "", private = TRUE,
  auto_init = FALSE, gitignore_template = "R")
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{name}{Character. One or more GitHub username or team name.}

\item{prefix}{Character. Common repository name prefix}

\item{suffix}{Character. Common repository name suffix}

\item{private}{Logical. Create private repositories?}

\item{auto_init}{Logical. Initialize the repository with a README.md?}

\item{gitignore_template}{Character. .gitignore template language.}
}
\description{
\code{create_repo} creates either individual or team repositories for a given
assignment.
}
\examples{
\dontrun{
create_repo("ghclass-test", c("user01","user02"), prefix = "hw01-")
}

}
\seealso{
Other github organization related functions: \code{\link{get_member}},
  \code{\link{get_pending_member}}, \code{\link{get_repo}},
  \code{\link{get_team_repos}}, \code{\link{invite_user}}
}
\concept{github organization related functions}
