% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{create_team}
\alias{create_team}
\title{Create team(s)}
\usage{
create_team(org, team, privacy = c("closed", "secret"))
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{team}{character, listing one or more teams}

\item{privacy}{character, level of privacy of teams, closed (visible to all
members of the organization) or secret (only visible to organization owners
and members of a team), default is closed}
}
\description{
\code{create_team} creates teams in your GitHub organization
}
\examples{
\dontrun{
create_team("ghclass",c("team01","team01"))
}

}
\seealso{
Other github organization team related functions: \code{\link{add_team_member}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_team_members}}, \code{\link{get_teams}},
  \code{\link{rename_team}}, \code{\link{team_delete}}
}
\concept{github organization team related functions}
