% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{fix_repo_name}
\alias{fix_repo_name}
\title{Fix repository names}
\usage{
fix_repo_name(repo_name)
}
\arguments{
\item{repo_name}{Character. Name of repository. Can be vector of list of characters.}
}
\value{
A character vector of repository names in the corrected format.
}
\description{
\code{fix_repo_name} replaces spaces in repository names with \code{_}. It also replaces non-alphanumeric characters and special characters other than \code{_}, \code{.}, or \code{-} with \code{-}.
}
\examples{
\dontrun{
fix_repo_name("base hw1")
fix_repo_name(c("base hw1", "final*draft"))
fix_repo_name(list("base hw1", "final*draft"))
}

}
