% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_admin}
\alias{get_admin}
\title{List repository administrators}
\usage{
get_admin(org)
}
\arguments{
\item{org}{Character. Name of a GitHub organization.}
}
\value{
A character vector of repository administrators.
}
\description{
List repository administrators
}
\examples{
\dontrun{
get_admin("Sta523-Fa17")
}

}
