% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_admins}
\alias{get_admins}
\title{List repository administrators}
\usage{
get_admins(org, verbose = FALSE)
}
\arguments{
\item{org}{Character. Name of GitHub organization.}

\item{verbose}{Logical. Display verbose output.

\examples
\dontrun{
get_admins("Sta523-Fa17")
}}
}
\value{
A list containing a character vector of repository administrators.
}
\description{
\code{get_admins} creates a list of repository administrators.
}
