% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_collaborator}
\alias{get_collaborator}
\title{List collaborator(s)}
\usage{
get_collaborator(repo, include_admin = TRUE, verbose = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{include_admin}{Logical. If FALSE, user names of users with Admin rights are not included, defaults to TRUE.}

\item{verbose}{Logical. Display verbose output.}
}
\value{
A list containing a character vector of user names.
}
\description{
\code{get_collaborator} Returns a vector of collaborator user names. Users with Admin rights are by default excluded, but can be included manually.
}
\examples{
\dontrun{
get_collaborators("Sta523-Fa17")
}

}
