% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_collaborators}
\alias{get_collaborators}
\title{List collaborators}
\usage{
get_collaborators(repo, include_admins = TRUE, verbose = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{include_admins}{Logical. If FALSE, user names of users with Admin rights are not included. Default is TRUE.}

\item{verbose}{Logical. Display verbose output.}
}
\value{
A list containing a character vector of user names.
}
\description{
\code{get_collaborators} Returns a vector of collaborator user names. Users with Admin rights are by default excluded, but can be included manually.
}
\examples{
\dontrun{
get_collaborators("Sta523-Fa17")
}

}
