% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{get_file}
\alias{get_file}
\title{Low level function for retrieving a file from a GitHub Repository}
\usage{
get_file(repo, file, branch = "master")
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{file}{Characer. Path to the file within the repository.}

\item{branch}{Character. Name of branch to use, defaults to "master".}
}
\description{
Low level function for retrieving a file from a GitHub Repository
}
\seealso{
Other file functions: \code{\link{add_content}},
  \code{\link{add_file}}, \code{\link{get_readme}},
  \code{\link{put_file}}
}
\concept{file functions}
