% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_member}
\alias{get_member}
\title{Get organization member}
\usage{
get_member(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression be excluded or included.}
}
\description{
\code{get_member} returns a (filtered) vector of organization memebers.
}
\examples{
\dontrun{
get_member("ghclass")
}

}
\seealso{
Other github organization related functions: \code{\link{get_pending_member}},
  \code{\link{get_repo}}, \code{\link{get_team_repos}},
  \code{\link{invite_user}}, \code{\link{org_remove}}
}
\concept{github organization related functions}
