% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_members}
\alias{get_members}
\title{Get organization members}
\usage{
get_members(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{filter}{character, regex pattern for matching (or excluding) repos.}

\item{exclude}{logical, should entries matching the regex be excluded or included.}
}
\description{
\code{get_members} returns a (filtered) vector of organization memebers.
}
\examples{
\dontrun{
get_members("ghclass")
}

}
\seealso{
Other github organization related functions: \code{\link{get_pending_members}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_repos}}, \code{\link{get_team_members}},
  \code{\link{get_team_repos}}, \code{\link{get_teams}}
}
