% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_pending_member}
\alias{get_pending_member}
\title{Get pending organization members}
\usage{
get_pending_member(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repos.}

\item{exclude}{Logical. Should entries matching the regular expression be excluded or included.}
}
\description{
\code{get_pending_member} returns a (filtered) vector of pending organization memebers.
}
\examples{
\dontrun{
get_pending_member("ghclass")
}

}
\seealso{
Other github organization related functions: \code{\link{create_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_members-deprecated}},
  \code{\link{get_repos-deprecated}},
  \code{\link{get_repo}}, \code{\link{get_team_repos}},
  \code{\link{invite_user}}
}
\concept{github organization related functions}
