% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_pending_members-deprecated}
\alias{get_pending_members-deprecated}
\alias{get_pending_members}
\title{Get pending organization members}
\usage{
get_pending_members(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{filter}{character, regex pattern for matching (or excluding) repos.}

\item{exclude}{logical, should entries matching the regex be excluded or included.}
}
\description{
\code{get_pending_members} returns a (filtered) vector of pending organization memebers.
}
\section{\code{get_pending_members}}{

For \code{get_pending_members}, use \code{\link{get_pending_member}}.
}

\examples{
\dontrun{
get_pending_members("ghclass")
}

}
\seealso{
\code{\link{ourPkg-deprecated}}

Other github organization related functions: \code{\link{create_individual_repo}},
  \code{\link{create_team_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_member}},
  \code{\link{get_repos-deprecated}},
  \code{\link{get_repo}}, \code{\link{get_team_repos}},
  \code{\link{invite_user}}
}
\concept{github organization related functions}
\keyword{internal}
