% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{get_pending_team_members}
\alias{get_pending_team_members}
\title{Get pending team members}
\usage{
get_pending_team_members(org, team = get_teams(org))
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{team}{character or data frame, listing one or more team}
}
\description{
\code{get_pending_team_members} returns a data frame of pending team members.
}
\examples{
\dontrun{
get_pending_team_members("ghclass",c("team01","team02"))
}

}
\seealso{
Other github organization team related functions: \code{\link{add_team_member}},
  \code{\link{create_team}},
  \code{\link{get_team_members}}, \code{\link{get_teams}},
  \code{\link{rename_team}}
}
\concept{github organization team related functions}
