% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{get_readme}
\alias{get_readme}
\title{Low level function for retrieving the README of a GitHub Repository}
\usage{
get_readme(repo, branch = "master")
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{branch}{Character. Name of branch to use, defaults to "master".}
}
\description{
Low level function for retrieving the README of a GitHub Repository
}
\seealso{
Other file functions: \code{\link{add_content}},
  \code{\link{add_file}}, \code{\link{get_file}},
  \code{\link{put_file}}
}
\concept{file functions}
