% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_repo}
\alias{get_repo}
\title{Get organization repository}
\usage{
get_repo(org, filter = NULL, exclude = FALSE, full_repo = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression in \code{filter} be excluded or included?}

\item{full_repo}{Logical. Should the full repository address be returned (e.g. \code{owner/name} instead of just \code{repo})?}
}
\description{
\code{get_repo} returns a (filtered) vector of repositories belonging to a GitHub organization.
}
\examples{
\dontrun{
get_repo("ghclass")
get_repo("ghclass", "hw1-")
}

}
\seealso{
Other github organization related functions: \code{\link{create_individual_repo}},
  \code{\link{create_team_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_members-deprecated}},
  \code{\link{get_pending_member}},
  \code{\link{get_repos-deprecated}},
  \code{\link{get_team_repos}}, \code{\link{invite_user}}
}
\concept{github organization related functions}
