% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_repo_collaborators-deprecated}
\alias{get_repo_collaborators-deprecated}
\alias{get_repo_collaborators}
\title{List repository collaborators}
\usage{
get_repo_collaborators(repo)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}
}
\value{
Character vector of collaborator usernames.
}
\description{
\code{get_repo_collaborators} returns collaborator usernames.
}
\section{\code{get_repo_collaborators}}{

For \code{get_repo_collaborators}, use \code{\link{get_collaborators}}.
}

\examples{
\dontrun{
get_repo_collaborators("Sta523-Fa17/hw1")
}


}
\seealso{
\code{\link{ourPkg-deprecated}}
}
\keyword{internal}
