% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_repos-deprecated}
\alias{get_repos-deprecated}
\alias{get_repos}
\title{Get organization repos}
\usage{
get_repos(org, filter = NULL, exclude = FALSE, full_repo = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression in \code{filter} be excluded or included?}

\item{full_repo}{Logical. Should the full repository name be returned (e.g. \code{org/repo} instead of just \code{repo})?}
}
\description{
\code{get_repos} returns a (filtered) vector of repos belonging to a GitHub organization.
}
\section{\code{get_repos}}{

For \code{get_repos}, use \code{\link{get_repo}}.
}

\examples{
\dontrun{
get_repos("ghclass")
get_repos("ghclass", "hw1-")
}

}
\seealso{
\code{\link{ourPkg-deprecated}}

Other github organization related functions: \code{\link{create_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_members-deprecated}},
  \code{\link{get_pending_member}}, \code{\link{get_repo}},
  \code{\link{get_team_repos}}, \code{\link{invite_user}}
}
\concept{github organization related functions}
\keyword{internal}
