% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_repos}
\alias{get_repos}
\title{Get organization repos}
\usage{
get_repos(org, filter = NULL, exclude = FALSE, full_repo = TRUE)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{filter}{character, regex pattern for matching (or excluding) repos.}

\item{exclude}{logical, should entries matching the regex be excluded or included.}

\item{full_repo}{logical, should the full repo name be returned (e.g. \code{org/repo} instead of just \code{repo})}
}
\description{
\code{get_repos} returns a (filtered) vector of repos belonging to a GitHub organization.
}
\examples{
\dontrun{
get_repos("ghclass")
get_repos("ghclass", "hw1-")
}

}
\seealso{
Other github organization related functions: \code{\link{get_members}},
  \code{\link{get_pending_members}},
  \code{\link{get_team_members}},
  \code{\link{get_team_repos}}, \code{\link{get_teams}}
}
