% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{get_team_members}
\alias{get_team_members}
\title{Get team members}
\usage{
get_team_members(org, team = get_teams(org), get_pending = TRUE)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{team}{character or data frame, listing one or more team}

\item{get_pending}{include pending team members}
}
\description{
\code{get_team_members} returns a data frame of teams and their members.
}
\examples{
\dontrun{
get_team_members("ghclass",c("team01","team02"))
}

}
\seealso{
Other github organization team related functions: \code{\link{add_team_member}},
  \code{\link{create_team}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_teams}}, \code{\link{rename_team}},
  \code{\link{team_delete}}
}
\concept{github organization team related functions}
