% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{get_team_repos}
\alias{get_team_repos}
\title{Get teams' repos}
\usage{
get_team_repos(org, team)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{team}{character or data frame, listing one or more team}
}
\description{
\code{get_team_repos} returns a (filtered) data frame of teams and their repos.
}
\examples{
\dontrun{
get_team_repos("ghclass",c("team01","team02"))
}

}
\seealso{
Other github organization related functions: \code{\link{create_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_members-deprecated}},
  \code{\link{get_pending_member}},
  \code{\link{get_repos-deprecated}},
  \code{\link{get_repo}}, \code{\link{invite_user}}
}
\concept{github organization related functions}
