% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watching.R
\name{get_watching}
\alias{get_watching}
\title{Get repos user is watching}
\usage{
get_watching(filter = NULL, exclude = FALSE)
}
\arguments{
\item{filter}{character, regex pattern for matching (or excluding) repositories.}

\item{exclude}{logical, should entries matching the regex be excluded or included.}
}
\description{
Returns all of the authenticated user's watched repositories. This should
match the list at https://github.com/watching. The function can also
filter the results for matching (or excluding) repositories.
}
\examples{
\dontrun{
get_watching()
get_watching("Sta523-Fa15")
}

}
\seealso{
Other github repo related functions: \code{\link{check_repo}},
  \code{\link{unwatch_repo}}
}
\concept{github repo related functions}
