% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{get_wercker_token}
\alias{get_wercker_token}
\title{Get wercker token}
\usage{
get_wercker_token()
}
\description{
\code{get_wercker_token} obtains the user's wercker authentication token.
}
\details{
This function looks for the token in the following places (in order):
\enumerate{
  \item Value of \code{wercker_token} variable in \code{.ghclass} environment
  (this is where the package caches the token).

\item Value of \code{WERCKER_TOKEN} environmental variable.

\item Contents of \code{~/.wercker/token} file.
}
}
\examples{
\dontrun{
get_wercker_token()
}

}
