% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghclass.R
\docType{package}
\name{ghclass}
\alias{ghclass}
\alias{ghclass-package}
\title{ghclass: A package for github based assignment management}
\description{
ghclass: A package for github based assignment management
}
\section{repository functions}{


\itemize{
  \item \code{One} -
  \item \code{Two} -
}
}

\section{organization functions}{


\itemize{
  \item \code{One} -
  \item \code{Two} -
}
}

\section{wercker functions}{


\itemize{
  \item \code{One} -
  \item \code{Two} -
}
}

