% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_github.R
\name{ghclass_api_v3_req}
\alias{ghclass_api_v3_req}
\title{Internal GitHub API helper}
\usage{
ghclass_api_v3_req(
  endpoint,
  ...,
  .send_headers = character(),
  version = "2022-11-28",
  limit = github_get_api_limit()
)
}
\description{
Internal GitHub API helper
}
\keyword{internal}
