% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{github_api_get_repos-deprecated}
\alias{github_api_get_repos-deprecated}
\alias{github_api_get_repos}
\title{Get repositories}
\usage{
github_api_get_repos(org)
}
\description{
Get repositories
}
\section{\code{github_api_get_repos}}{

For \code{github_api_get_repos}, use \code{\link{github_api_get_repo}}.
}

\seealso{
\code{\link{ourPkg-deprecated}}
}
\keyword{internal}
