% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_get_file.R
\name{github_api_repo_get_file}
\alias{github_api_repo_get_file}
\title{Internal GitHub API helper}
\usage{
github_api_repo_get_file(repo, path, branch = NULL)
}
\description{
Internal GitHub API helper
}
\keyword{internal}
