% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_get_token.R
\name{github_get_token}
\alias{github_get_token}
\title{Get github token}
\usage{
github_get_token()
}
\description{
`github_get_token`` obtains the user's github authentication token.
}
\details{
This function looks for the token in the following places (in order):
\itemize{
\item Value of \code{GITHUB_PAT} environmental variable.
\item Value of \code{GITHUB_TOKEN} environmental variable.
\item Contents of \code{~/.github/token} file.
}
}
\examples{
\dontrun{
github_get_token()
}

}
\seealso{
Other authentication: \code{\link{github_reset_token}},
  \code{\link{github_set_token}},
  \code{\link{github_test_token}}, \code{\link{with_pat}}
}
\concept{authentication}
