% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_whoami.R
\name{github_orgs}
\alias{github_orgs}
\title{Returns a tibble with details on the authenticated user's (based on the current PAT)
GitHub organization memberships.}
\usage{
github_orgs(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Should status messages be shown.}
}
\description{
Returns a tibble with details on the authenticated user's (based on the current PAT)
GitHub organization memberships.
}
\examples{
\dontrun{
github_orgs()
}

}
