% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_reset_token.R
\name{github_reset_token}
\alias{github_reset_token}
\title{Reset GitHub token}
\usage{
github_reset_token()
}
\description{
\code{github_reset_token} removes any value stored in the \code{GITHUB_PAT}
enivronmental variable.
}
\examples{
\dontrun{
token = "0123456789ABCDEF0123456789ABCDEF01234567"
github_set_token("0123456789ABCDEF0123456789ABCDEF01234567")
get_github_token() == token
github_reset_token()
get_github_token() == token
}

}
\seealso{
Other authentication: \code{\link{get_github_token}},
  \code{\link{github_set_token}},
  \code{\link{test_github_token}}, \code{\link{with_pat}}
}
\concept{authentication}
