% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_sitrep.R
\name{github_sitrep}
\alias{github_sitrep}
\title{Reports on the current GitHub configuration and access privledges. See more details on scopes
\href{https://developer.github.com/apps/building-oauth-apps/understanding-scopes-for-oauth-apps/}{here}.}
\usage{
github_sitrep()
}
\description{
Reports on the current GitHub configuration and access privledges. See more details on scopes
\href{https://developer.github.com/apps/building-oauth-apps/understanding-scopes-for-oauth-apps/}{here}.
}
\examples{
\dontrun{
github_sitrep()
}

}
