% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_test_token.R
\name{github_test_token}
\alias{github_test_token}
\alias{test_github_token}
\title{Test github token}
\usage{
github_test_token(token = github_get_token())
}
\arguments{
\item{token}{github api personal access token}
}
\description{
github_test_token` checks if a token is valid by attempting to authenticate
with the GitHub api.
}
\examples{
\dontrun{
github_test_token()
github_test_token("bad_token")
}

}
