% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_token_scopes.R
\name{github_token_scopes}
\alias{github_token_scopes}
\title{Returns a vector of the o-auth scopes granted to the current PAT.}
\usage{
github_token_scopes(quiet = FALSE)
}
\description{
Returns a vector of the o-auth scopes granted to the current PAT.
}
\examples{
\dontrun{
github_token_scopes()
}

}
