% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_whoami.R
\name{github_whoami}
\alias{github_whoami}
\title{Get github login}
\usage{
github_whoami()
}
\description{
\code{github_whoami} returns the login of the authenticated user.
}
\examples{
\dontrun{
github_whoami()
}

}
