% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_create.R
\name{issue_create}
\alias{issue_create}
\title{Create an issue}
\usage{
issue_create(repo, title, body, labels = character(), assignees = character())
}
\arguments{
\item{repo}{Character. Address of one or more repositories in \code{owner/name} format.}

\item{title}{Character. Title of the issue.}

\item{body}{Character. Content of the issue.}

\item{labels}{Character. Vector of the labels to associate with this issue}

\item{assignees}{Character. Vector of logins for users assigned to the issue.}
}
\description{
\code{issue_create} creates an issue for a GitHub repository.
}
