% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_close.R
\name{issue_open}
\alias{issue_open}
\title{Open an issue}
\usage{
issue_open(repo, number)
}
\arguments{
\item{repo}{Character. Address of one or more repositories in \code{owner/name} format.}

\item{number}{Integer. Issue number of the issue to close.}
}
\description{
\code{issue_open} creates an issue for a GitHub repository.
}
