% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo_clone.R
\name{local_repo_clone}
\alias{local_repo_clone}
\alias{repo_clone}
\title{Local repository tools}
\usage{
local_repo_clone(repo, local_path = "./", branch = "master",
  git = require_git(), options = character(), verbose = FALSE)
}
\arguments{
\item{repo}{GitHub repo names with the form \code{owner/name}.}

\item{local_path}{Local directory to store cloned repos.}

\item{branch}{Repository branch to use.}

\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your `PATH``, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}
}
\description{
Clones repositories from GitHub to a local directory.
}
\examples{
\dontrun{
g = org_repos("Sta323-Sp18","hw3-")
local_repo_clone(g, "hw3")
}

}
