% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo_log.R
\name{local_repo_log}
\alias{local_repo_log}
\title{Local repository tools}
\usage{
local_repo_log(repo_dir, max = 100)
}
\arguments{
\item{repo_dir}{Vector of repo directories or a single directory containing one or more repos.}

\item{max}{Maximum number of log entries to retrieve per repo.}
}
\description{
Returns up to \code{max} entries from a git repo's log
}
\examples{
\dontrun{
local_repo_log("hw1/")
}

}
