% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{mirror_repo}
\alias{mirror_repo}
\title{Mirror repository}
\usage{
mirror_repo(source_repo, target_repo, verbose = FALSE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{target_repo}{Character. One or more repository addresses in "owner/name" format.}

\item{verbose}{Logical. Display verbose output.}
}
\description{
\code{mirror_repo} mirrors the content of one repository to another repository, or set of
repositories.
}
\examples{
\dontrun{
mirror_repo("ghclass-test/hw1", c("ghclass-test/hw1-Team1", "ghclass-test/hw1-Team2"))
mirror_repo("ghclass-test/hw1", org_repos("ghclass-test","hw1-"))
}

}
