% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_admins.R
\name{org_admins}
\alias{org_admins}
\title{List repository administrators}
\usage{
org_admins(org)
}
\arguments{
\item{org}{Character. Name of a GitHub organization.}
}
\value{
A character vector of repository administrators.
}
\description{
List repository administrators
}
\examples{
\dontrun{
org_admins("ghclass-test")
org_admins("rundel")
}

}
