% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_exists.R
\name{org_exists}
\alias{org_exists}
\title{Check if organization exists}
\usage{
org_exists(org)
}
\arguments{
\item{org}{Character. Organization name to be checked.}
}
\value{
Vector of logical values
}
\description{
\code{org_exists} returns TRUE if the supplied org(s) exist on GitHub and FALSE otherwise.
}
\examples{
\dontrun{
org_exists(c("rundel","ghclass-demo"))
}

}
