% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_invite.R
\name{org_invite}
\alias{org_invite}
\title{Invite user(s)}
\usage{
org_invite(org, user)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{user}{Character, character vector, or list. Listing one or more user names.}
}
\description{
\code{org_invite} invites user(s) to your organization.
}
\examples{
\dontrun{
user = c("Alice","Bob","Carol","Dave","Eve")
org_invite("Sta523-Fa17", user)
}

}
