% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_pending_members.R
\name{org_pending_members}
\alias{org_pending_members}
\title{Get pending organization members}
\usage{
org_pending_members(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repos.}

\item{exclude}{Logical. Should entries matching the regular expression be excluded or included.}
}
\description{
\code{org_pending_members} returns a (filtered) vector of pending organization memebers.
}
\examples{
\dontrun{
org_pending_members("ghclass")
}

}
\seealso{
Other github organization related functions: \code{\link{org_invite}},
  \code{\link{org_members}}, \code{\link{org_repos}},
  \code{\link{team_repos}}
}
\concept{github organization related functions}
