% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_repo_stats.R
\name{org_repo_stats}
\alias{org_repo_stats}
\title{Get organization repository}
\usage{
org_repo_stats(org)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}
}
\description{
\code{org_repo_stats} returns a tibble of repositories belonging to a GitHub organization along with some
basic statistics about those repositories.
}
\examples{
\dontrun{
org_repo_stats("ghclass")
}

}
