% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_teams.R
\name{org_teams}
\alias{org_teams}
\title{Get organization teams}
\usage{
org_teams(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{filter}{character, regex pattern for matching (or excluding) teams}

\item{exclude}{logical, should entries matching the regex be excluded or included.}
}
\description{
\code{org_teams} returns a (filtered) vector of teams in the organization.
}
