% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_teams_v4.R
\name{org_teams_v4}
\alias{org_teams_v4}
\title{Get organization teams}
\usage{
org_teams_v4(org)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}
}
\description{
Get organization teams
}
\examples{
\dontrun{
org_teams_v4("ghclass-demo")
}

}
