% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_add_file}
\alias{peer_add_file}
\title{Add files to repositories based on roster}
\usage{
peer_add_file(org, roster, to = c("r", "a"), file, dblind = FALSE,
  prefix = "", suffix = "", message = NULL, branch = "master",
  overwrite = FALSE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{r*} columns that specify review assignments as values of the vector \code{user_random}. See \code{peer_create_rform}.}

\item{to}{Character. Specifies whether the file is to be added to reviewers' (\code{r}) or authors' (\code{a}) folders. If \code{to = "r"}, the function places the file into each of the folders containing the anonymized author id on reviewers' repositores. If \code{to = "a"}, the function places the file into each of the folders containing the reviewers' files on authors' repositories.}

\item{file}{Character. File name of file to be added.}

\item{dblind}{Logical. Specifies whether review is conducted double-blind (i.e. neither reviewer nor author can identify each other), or single-blind (i.e. authors remain anonymous but reviewer identities are revealed). If \code{dblind = TRUE}, reviewer folders are identified by the reviewer's ID. If \code{dblind = FALSE}, reviewer folders are identified by the original user names. Defaults to \code{FALSE}.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch the file should be committed to, defaults to \code{master}.}

\item{overwrite}{Logical. Whether existing files in reviewers' repositories should be overwritten, defaults to \code{FALSE}.}
}
\description{
\code{peer_add_file()} is the peer review version of \code{repo_add_file()}. It takes a local file and adds it to author- or reviewer-specific folders in students' repositories based on the peer review roster. The function's main purpose is to distribute feedback forms into the folders containing copies of authors' or reviewers' files.
}
