% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_add_file_aut}
\alias{peer_add_file_aut}
\title{Add local files to reviewer-specific folders on authors' repositories}
\usage{
peer_add_file_aut(org, roster, local_path, double_blind = FALSE,
  prefix = "", suffix = "", message = NULL, branch = "master",
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{rev*} columns that specify review assignments as values of the vector \code{user_random}.}

\item{local_path}{Character. File name of file to be added.}

\item{double_blind}{Logical. Specifies whether review is conducted double-blind (i.e. neither reviewer nor author can identify each other), or single-blind (i.e. authors remain anonymous but reviewer identities are revealed). If \code{double_blind = TRUE}, reviewer folders are identified by the reviewer's ID. If \code{double_blind = FALSE}, reviewer folders are identified by the original user names. Defaults to \code{FALSE}.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch the file should be committed to, defaults to \code{master}.}

\item{overwrite}{Logical. Whether existing files in reviewers' repositories should be overwritten, defaults to \code{FALSE}.}

\item{verbose}{Logical. Should success / failure messages be printed.}
}
\description{
\code{peer_add_file_aut()} takes a local file and adds it to reviewer-specific folders on authors' repositories. The function's main purpose is to distribute rating forms into the reviewer-specific folders on authors' repositories.
}
