% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_add_file_rev}
\alias{peer_add_file_rev}
\title{Add local files to author-specific folders on reviewers' review repositories}
\usage{
peer_add_file_rev(org, roster, local_path, prefix = "", suffix = "",
  message = NULL, branch = "master", overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{rev*} columns that specify review assignments as values of the vector \code{user_random}.}

\item{local_path}{Character. File name of file to be added.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch the file should be committed to, defaults to \code{master}.}

\item{overwrite}{Logical. Whether existing files in reviewers' repositories should be overwritten, defaults to \code{FALSE}.}

\item{verbose}{Logical. Should success / failure messages be printed.}
}
\description{
\code{peer_add_file_rev()} takes a local file and adds it to author-specific folders on reviewers' repositories. The function's main purpose is to distribute review forms into the correct author-specific folders on reviewers' repositories.
}
