% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_anonymize_file}
\alias{peer_anonymize_file}
\title{Anonymize file}
\usage{
peer_anonymize_file(file)
}
\arguments{
\item{file}{Character. Local file path.}
}
\description{
\code{anonymize_file} removes all identifying student information
}
