% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_blankfeedback}
\alias{peer_blankfeedback}
\title{Create feedback form}
\usage{
peer_blankfeedback(n, title = "Feedback form", file = "feedback_blank",
  output = "github_document", write_rmd = TRUE)
}
\arguments{
\item{n}{Numerical. Number of grade fields to be included in .Rmd YAML.}

\item{title}{Character. Title of form, defaults to "Feedback form."}

\item{file}{Character. File name of RMarkdown document, defaults to \code{feedback_blank}.}

\item{output}{Character. Output parameter for \code{.Rmd} file, defaults to \code{github_document}.}

\item{write_rmd}{Logical. Whether the feedback form should be saved to a \code{.Rmd} file in the current working directory, defaults to TRUE.}
}
\description{
Create feedback form
}
