% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_collect_grade}
\alias{peer_collect_grade}
\title{Extract grades from feedback forms}
\usage{
peer_collect_grade(org, prefix = "", suffix = "", file, roster,
  write_csv = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{file}{Character. File name of feedback form (must be .Rmd document).}

\item{roster}{Character. File path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{r*} columns that specify review assignments as values of the vector \code{user_random}. See \code{peer_create_feedback}.}

\item{write_csv}{Logical. Whether the roster data frame should be saved to a \code{.csv} file in the current working directory, defaults to TRUE.}

\item{name}{Character. One or more GitHub user or team name(s).}
}
\description{
The \code{peer_collect_grade()} function collects grade information from the YAML of a feedback form within a student's repository. It outputs a new .csv file, with rows specifying individual question grades for each student.
}
