% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_create_aform}
\alias{peer_create_aform}
\title{Create author feedback form}
\usage{
peer_create_aform(category = c("helpfulness", "accuracy", "fairness"),
  title = "Author feedback form", file = "afeedback_blank",
  output = "github_document", write_rmd = TRUE, overwrite = FALSE)
}
\arguments{
\item{category}{Character. Categories to be included in the feedback form, defaults to \code{c("helpfulness", "accuracy", "fairness")}.}

\item{title}{Character. Title of form, defaults to "Author feedback form."}

\item{file}{Character. File name of RMarkdown document, defaults to \code{rfeedback_blank}.}

\item{output}{Character. Output parameter for \code{.Rmd} file, defaults to \code{github_document}.}

\item{write_rmd}{Logical. Whether the feedback form should be saved to a \code{.Rmd} file in the current working directory, defaults to TRUE.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to \code{FALSE}.}
}
\description{
\code{peer_create_aform} creates a short feedback form for authors to rate the feedback they got from reviewers.
}
