% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_create_form_review}
\alias{peer_create_form_review}
\title{Create reviewer feedback form}
\usage{
peer_create_form_review(n, title = "Reviewer feedback form",
  fname = "feedback_blank_review", output = "github_document",
  write_rmd = TRUE, overwrite = FALSE, dblind = FALSE)
}
\arguments{
\item{n}{Numerical. Number of score fields to be included in .Rmd YAML.}

\item{title}{Character. Title of form, defaults to "Reviewer feedback form."}

\item{fname}{Character. File name of RMarkdown document to be written to memory, defaults to \code{feedback_blank_review}.}

\item{output}{Character. Output parameter for \code{.Rmd} file, defaults to \code{github_document}.}

\item{write_rmd}{Logical. Whether the feedback form should be saved to a \code{.Rmd} file in the current working directory, defaults to \code{TRUE}.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to \code{FALSE}.}

\item{dblind}{Logical. If \code{dblind = TRUE}, the YAML will contain an \code{author} field, defaults to \code{FALSE}.}
}
\description{
\code{peer_create_form_review} creates blank feedback forms for reviewers based on the user-specified number of questions.
}
