% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_peer.R
\name{peer_mirror_original}
\alias{peer_mirror_original}
\title{Mirror original file(s)}
\usage{
peer_mirror_original(source_repo, path, target_folder, message = NULL,
  branch = "master", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{path}{Character or character vector. Name(s) of file(s) to be moved.}

\item{target_folder}{Character. Name of folder containing file on \code{target_repo}.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch to use, defaults to \code{master}.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to \code{FALSE}.}

\item{verbose}{Logical. Should success/failure messages be printed, defaults to \code{TRUE}.}
}
\description{
\code{peer_mirror_original} mirrors original file(s) from an author's repository into a reviewer-specific folder on the author's repository and returns a list of contents.
}
\seealso{
Other peer review functions: \code{\link{peer_assign}},
  \code{\link{peer_file_add_aut}},
  \code{\link{peer_file_add_rev}},
  \code{\link{peer_form_create_rating}},
  \code{\link{peer_form_create_review}},
  \code{\link{peer_init}},
  \code{\link{peer_mirror_review}},
  \code{\link{peer_return}},
  \code{\link{peer_roster_create}},
  \code{\link{peer_score_rating}},
  \code{\link{peer_score_review}}
}
\concept{peer review functions}
