% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_peer.R
\name{peer_mirror_review}
\alias{peer_mirror_review}
\title{Mirror reviewer file(s)}
\usage{
peer_mirror_review(content_og, source_repo, target_repo, path,
  form_review = NULL, source_folder = NULL, target_folder = NULL,
  message = NULL, branch = "master", verbose = TRUE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{target_repo}{Character. Address of repository in "owner/name" format.}

\item{path}{Character or character vector. Name(s) of file(s) to be moved.}

\item{form_review}{Character. File name of reviewer feedback form (must be .Rmd document). If \code{NULL}, no review form will be moved.}

\item{source_folder}{Character. Name of folder containing file on \code{source_repo}.}

\item{target_folder}{Character. Name of folder containing file on \code{target_repo}.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch to use, defaults to \code{master}.}

\item{verbose}{Logical. Should success/failure messages be printed, defaults to \code{TRUE}.}

\item{og_content}{Character. List of files with the content of an author's original assignment file(s)}
}
\description{
\code{peer_mirror_review} mirrors user-specified file(s) (and review forms if applicable) from a reviewer's review repository into a reviewer-specific folder on the author's repository and returns a list of URLs. The function conducts a check whether author assignment files were changed by the reviewer; the function returns \code{NULL} if there is no change.
}
\seealso{
Other peer review functions: \code{\link{peer_assign}},
  \code{\link{peer_file_add_aut}},
  \code{\link{peer_file_add_rev}},
  \code{\link{peer_form_create_rating}},
  \code{\link{peer_form_create_review}},
  \code{\link{peer_init}},
  \code{\link{peer_mirror_original}},
  \code{\link{peer_return}},
  \code{\link{peer_roster_create}},
  \code{\link{peer_score_rating}},
  \code{\link{peer_score_review}}
}
\concept{peer review functions}
