% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_return}
\alias{peer_return}
\title{Return peer feedback to authors}
\usage{
peer_return(org, roster, file, dblind = FALSE, prefix = "",
  suffix = "", message = "Returning review", branch = "master",
  overwrite = FALSE)
}
\arguments{
\item{org}{Character. Name of GitHub Organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{r*} columns that specify review assignments as values of the vector \code{user_random}. See \code{peer_create_feedback}.}

\item{file}{Character. File name or vector of file names to be included.}

\item{dblind}{Logical. Specifies whether review is conducted double-blind (i.e. neither reviewer nor author can identify each other), or single-blind (i.e. authors remain anonymous but reviewer identities are revealed). If \code{dblind = TRUE}, reviewer folders are identified by the anonymized user IDs in the roster's \code{user_random} column. If \code{dblind = FALSE}, reviewer folders are identified by the original user names. Defaults to \code{FALSE}.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{message}{Character. Commit message, defaults to "Assigning review."}

\item{branch}{Character. Name of branch the file should be committed to, defaults to \code{master}.}

\item{overwrite}{Logical. Whether existing files in reviewers' repositories should be overwritten, defaults to \code{FALSE}.}
}
\description{
Return peer feedback to authors
}
