% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_roster}
\alias{peer_roster}
\title{Create peer review roster}
\usage{
peer_roster(m, user, seed = 12345, write_csv = TRUE)
}
\arguments{
\item{m}{Numeric. Number of reviews per user. Must be larger than zero and smaller than the number of users.}

\item{user}{Character. A vector of GitHub user names.}

\item{seed}{Numeric. Random seed for assignment, defaults to \code{12345}.}

\item{write_csv}{Logical. Whether the roster data frame should be saved to a \code{.csv} file in the current working directory, defaults to TRUE.}
}
\description{
\code{peer_roster} creates data frame of random assignments of author files to reviewers. By default, the output is saved to a \code{.csv} file in the current working directory that incorporates the current date and random seed as part of the file name.
}
