% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_roster_create}
\alias{peer_roster_create}
\title{Create peer review roster}
\usage{
peer_roster_create(n_rev, user, seed = 12345, write_csv = TRUE)
}
\arguments{
\item{n_rev}{Numeric. Number of reviews per user. Must be larger than zero and smaller than the number of users.}

\item{user}{Character. A vector of GitHub user names.}

\item{seed}{Numeric. Random seed for assignment, defaults to \code{12345}.}

\item{write_csv}{Logical. Whether the roster data frame should be saved to a \code{.csv} file in the current working directory, defaults to TRUE.}
}
\description{
\code{peer_roster_create} creates data frame of random assignments of author files to reviewers. By default, the output is saved to a \code{.csv} file in the current working directory that incorporates the current date and random seed as part of the file name.
}
\examples{
\dontrun{
peer_roster_create(3, c("anya-ghclass", "bruno-ghclass", "celine-ghclass", "diego-ghclass"))
}

}
\seealso{
Other peer review functions: \code{\link{peer_assign}},
  \code{\link{peer_file_add_aut}},
  \code{\link{peer_file_add_rev}},
  \code{\link{peer_form_create_rating}},
  \code{\link{peer_form_create_review}},
  \code{\link{peer_init}}, \code{\link{peer_return}},
  \code{\link{peer_score_rating}},
  \code{\link{peer_score_review}},
  \code{\link{repo_file_mirror}}
}
\concept{peer review functions}
