% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_score_rating}
\alias{peer_score_rating}
\title{Collect scores from rating forms}
\usage{
peer_score_rating(org, roster, form_rating, double_blind = FALSE,
  prefix = "", suffix = "", write_csv = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{rev*} columns that specify review assignments as values of the vector \code{user_random}.}

\item{form_rating}{Character. File name of rating feedback form (must be .Rmd document).}

\item{double_blind}{Logical. Specifies whether review is conducted double-blind (i.e. neither reviewer nor author can identify each other), or single-blind (i.e. authors remain anonymous but reviewer identities are revealed). If \code{double_blind = TRUE}, reviewer folders are identified by the anonymized user IDs in the roster's \code{user_random} column. If \code{double_blind = FALSE}, reviewer folders are identified by the original user names. Defaults to \code{FALSE}.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{write_csv}{Logical. Whether the roster data frame should be saved to a \code{.csv} file in the current working directory, defaults to TRUE.}
}
\description{
The \code{peer_score_rating()} function collects score information from the YAML of a rating form within authors' repositories. It outputs a new .csv file, with rows specifying individual question scores for each student.
}
