% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_create.R
\name{pr_create}
\alias{pr_create}
\title{Create pull request}
\usage{
pr_create(repo, title, base, head = "master", body = "", draft = TRUE)
}
\arguments{
\item{repo}{Character. Address of one or more repositories in "owner/name" format.}

\item{title}{Character. Title of the pull request.}

\item{base}{Character. The name of the branch where your changes are implemented. In creating a pull request from a
fork then use \code{username:branch} as the format.}

\item{head}{Character. The branch you want the changed pulled into.}

\item{body}{Character. The text contents of the pull request.}

\item{draft}{Logical. Should the pull request be created as a draft pull request (these cannot be merged
until allowed by the author).}
}
\description{
\code{pr_create} creates a pull request on GitHub from the \code{base} branch to the \code{head} branch.
}
