% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{protect_branch}
\alias{protect_branch}
\title{Protect branch}
\usage{
protect_branch(repo, branch = "master")
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}

\item{cur_branch}{name of existing branch}

\item{new_branch}{name of branch to create}
}
\description{
\code{protect_branch} turns on protection for the specified branch. See
\link{https://help.github.com/en/articles/about-protected-branches} for more details
on what this changes.
}
\seealso{
Other branch functions: \code{\link{create_branch}},
  \code{\link{unprotect_branch}}
}
\concept{branch functions}
