% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo.R
\name{push_repos}
\alias{push_repos}
\alias{pull_repos}
\title{Push local repos via git}
\usage{
push_repos(repo_dirs, remotes = "origin master", git = require_git(),
  options = "", verbose = FALSE)
}
\arguments{
\item{repo_dirs}{Vector of repo directories or a single directory containing one or more repos.}

\item{remotes}{Vector of remote names or urls}

\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your \code{PATH}, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}
}
\description{
\code{push_repos} uses the git binary to push local changes to the provided remote.
}
\examples{
\dontrun{
update_repos("hw1/")
update_repos(c("hw1/team01","hw1/team02","hw1/team03"))
}

}
\seealso{
Other local repo functions: \code{\link{add_repos}},
  \code{\link{clone_repos}}, \code{\link{commit_repos}},
  \code{\link{update_repos}}
}
