% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{put_file}
\alias{put_file}
\title{Low level function for adding a file to a Github repository}
\usage{
put_file(repo, path, content, message = NULL, branch = "master",
  verbose = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{path}{Character. File path within the repository.}

\item{content}{Character or raw. Content of the file.}

\item{message}{Character. Commit message. If not provided, a custom character string will be created.}

\item{branch}{Character. Name of branch to use, defaults to "master".}

\item{verbose}{Logical. Should success / failure messages be printed}
}
\description{
Low level function for adding a file to a Github repository
}
\seealso{
Other file functions: \code{\link{add_content}},
  \code{\link{add_file}}
}
\concept{file functions}
