% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo_rename.R
\name{rename_local_repo}
\alias{rename_local_repo}
\title{Rename local directories using a vector of patterns and replacements}
\usage{
rename_local_repo(repo_dir, pattern, replacement)
}
\arguments{
\item{repo_dir}{Character. Vector of repo directories or a single directory containing one or more repos.}

\item{pattern}{Character. One or more regexp patterns to match to directory names.}

\item{replacement}{Character.  One or more text strings containing the replacement value for matched patterns.}
}
\description{
This is particularly helpful for renaming student repos to include
something more useful like \code{Last, First name} or \code{netid} so that
folder ordering matches student ordering within your LMS.
}
