% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{rename_repo}
\alias{rename_repo}
\title{Rename repository}
\usage{
rename_repo(repo, new_name)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{new_name}{Character. New name of repository in the "name" format.}
}
\description{
\code{rename_repo} renames repositories. Use with caution as repositories retain their
unique identifier upon renaming and can be accessed under their old names due to
GitHub re-directing.
}
\examples{
\dontrun{
rename_repo("ghclass-test/hw1", "homework1")
}

}
