% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{rename_team}
\alias{rename_team}
\title{Rename existing team(s)}
\usage{
rename_team(org, team, new_team)
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{team}{character, one or more existing team names}

\item{new_team}{character, one or more new team names}
}
\description{
\code{rename_team} renames an existing team within the given GitHub organization.
}
\examples{
\dontrun{
rename_team("ghclass-test", "hw1-team01", "hw01-team01")
}

}
\seealso{
Other github organization team related functions: \code{\link{add_team_member}},
  \code{\link{create_team}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_team_members}}, \code{\link{get_teams}},
  \code{\link{team_delete}}
}
\concept{github organization team related functions}
