% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_codeowner.R
\name{repo_add_codeowner}
\alias{repo_add_codeowner}
\title{Add CODEOWNERS to repo}
\usage{
repo_add_codeowner(repo, owner, message = NULL, overwrite = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{owner}{Character. Username of the owner. Without any prefix.}

\item{message}{Character. Commit message.}

\item{overwrite}{Logical. Should any existing CODEOWNERS file be replaced?}
}
\value{
Called for its side effects. invisibly return a list containing the
results of the relevant GitHub API calls.
}
\description{
Add / replace \code{.github/CODEOWNERS} files in a GitHub repository.
Note that due to delays in caching, files that have been added
very recently might not yet be displayed as existing and might accidentally
be overwritten. The \code{owner} will be responsible for all files in the repo.
}
\examples{
\dontrun{
repo_add_codeowner("orgname/example-repo", "dajmcdon")
}
}
