% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_team.R, R/repo_add_user.R
\name{repo_add_team}
\alias{repo_add_team}
\alias{repo_add_user}
\alias{add_user_to_repo}
\alias{add_team_to_repo}
\title{Add a user or team to a repository}
\usage{
repo_add_team(repo, team, permission = c("push", "pull", "admin"))

repo_add_user(repo, user, permission = c("push", "pull", "admin"))
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{team}{Character. One or more GitHub team names.}

\item{permission}{Character. Permission to be granted to team for repo ("push", "pull", or "admin"), defaults to "push".

\code{pull} results in read privileges, \code{push} in write privileges,
and \code{admin} in Admin privileges for the team in the respective repository.
Note that permissions will overwrite existing access privileges.}

\item{user}{Character. One or more GitHub usernames.}
}
\description{
Add a user or team to a repository
}
\examples{
\dontrun{
repo_add_user("ghclass-test/hw1-team1", c("user01", "user02"))

## Adding users to their individual repositories
user = c("user01", "user02")
repo_add_user(paste0("ghclass-test/hw2-", user), user)
}

}
