% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_randomized_file.R
\name{repo_add_randomized}
\alias{repo_add_randomized}
\alias{repo_add_randomized_text_file}
\alias{repo_add_sampled_data_file}
\title{Add randomized versions of a file (unique) to each repo}
\usage{
repo_add_randomized_text_file(
  repo,
  file,
  nreplace = 2,
  replace_str = "",
  message = NULL,
  repo_folder = NULL,
  branch = NULL,
  preserve_path = FALSE,
  overwrite = FALSE
)

repo_add_sampled_data_file(
  repo,
  file,
  n,
  replace = FALSE,
  message = NULL,
  repo_folder = NULL,
  branch = NULL,
  preserve_path = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{file}{Character. Local file path(s) of file or files to be added.}

\item{nreplace}{Integer. Number of lines to replace.}

\item{replace_str}{Character. String to replace previously existing lines.}

\item{message}{Character. Commit message.}

\item{repo_folder}{Character. Name of folder on repository to save the file(s) to. If the folder does not exist on the repository, it will be created.}

\item{branch}{Character. Name of branch to use.}

\item{preserve_path}{Logical. Should the local relative path be preserved.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to FALSE.}

\item{n}{number of items to sample}

\item{replace}{do we sample with replacement?}
}
\description{
\code{repo_add_randomized_text_file()} takes in text files (R code or a markdown document for example)
and randomly replaces some lines with a text string (only those lines which had text to begin with).
The result is uploaded to an individual repo. Thus, different repos will receive
different versions of the file.

\code{repo_add_sampled_data_file} takes an RDS file (or set of such) and resamples it
(optionally with replacement) before pushing to a repo.
}
