% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_team.R
\name{repo_add_team}
\alias{repo_add_team}
\title{Add a team to a repository}
\usage{
repo_add_team(repo, team, permission = c("push", "pull", "admin"))
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{team}{Character. Team name.}

\item{permission}{Character. Permission to be granted to team for repo ("push", "pull", or "admin"), defaults to "push".
\itemize{
\item pull - can pull, but not push to or administer this repository.
\item push - can pull and push, but not administer this repository.
\item admin - can pull, push and administer this repository.
}}
}
\description{
Add a team to a repository
}
