% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_user.R
\name{repo_add_user}
\alias{repo_add_user}
\title{Add a user to a repository}
\usage{
repo_add_user(repo, user, permission = c("push", "pull", "admin"))
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{user}{Character. One or more GitHub usernames.}

\item{permission}{Character. Permission to be granted to team for repo ("push", "pull", or "admin"), defaults to "push".
\itemize{
\item pull - can pull, but not push to or administer this repository.
\item push - can pull and push, but not administer this repository.
\item admin - can pull, push and administer this repository.
}}
}
\description{
Add a user to a repository
}
\examples{
\dontrun{
repo_add_user("ghclass-test/hw1-team1", c("user01", "user02"))

## Adding users to their individual repositories
user = c("user01", "user02")
repo_add_user(paste0("ghclass-test/hw2-", user), user)
}

}
