% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_clone_url.R
\name{repo_clone_url}
\alias{repo_clone_url}
\title{Get the cloning url for a GitHub repo}
\usage{
repo_clone_url(repo, type = c("https", "ssh"))
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{type}{Character. Either https or ssh.}
}
\value{
A character vector of urls
}
\description{
Get the cloning url for a GitHub repo
}
