% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_collaborators.R
\name{repo_collaborators}
\alias{repo_collaborators}
\title{List collaborator(s)}
\usage{
repo_collaborators(repo, include_admins = TRUE)
}
\arguments{
\item{repo}{Character. Address of one or more repositories in \code{owner/name} format.}

\item{include_admins}{Logical. If \code{FALSE}, user names of users with Admin rights are not included, defaults to \code{TRUE}.}
}
\value{
A tibble with two columns, \code{repo} and \code{username}.
}
\description{
\code{repo_collaborators} Returns a vector of collaborator user names. Users with admin rights are by default excluded.
}
\examples{
\dontrun{
repo_collaborators("ghclass-test/test2")
}

}
