% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_commits.R
\name{repo_commits}
\alias{repo_commits}
\title{Get repository commits}
\usage{
repo_commits(
  repo,
  branch = "master",
  sha = branch,
  path = NULL,
  author = NULL,
  since = NULL,
  until = NULL,
  quiet = FALSE
)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{branch}{Character.    Branch to list commits from.}

\item{sha}{Character.    SHA to start listing commits from.}

\item{path}{Character.    Only commits containing this file path will be returned.}

\item{author}{Character.    GitHub login or email address by which to filter commit author.}

\item{since}{Character.    Only commits after this date will be returned, expects \code{YYYY-MM-DDTHH:MM:SSZ} format.}

\item{until}{Character.    Only commits before this date will be returned, expects \code{YYYY-MM-DDTHH:MM:SSZ} format.}

\item{quiet}{Logical. Should an error message be printed if the repo does not exist.
Character.}
}
\description{
\code{repo_commits} returns a tibble of repositories belonging to a GitHub organization along with some
basic statistics about those repositories.
}
\examples{
\dontrun{
org_repo_stats("ghclass")
}

}
