% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_contributors.R
\name{repo_contributors}
\alias{repo_contributors}
\title{Get repository contributor(s)}
\usage{
repo_contributors(repo, include_admins = TRUE)
}
\arguments{
\item{repo}{Character. Address of one or more repositories in \code{owner/name} format.}

\item{include_admins}{Logical. If \code{FALSE}, usernames of users with Admin rights are excluded, defaults to \code{TRUE}.}
}
\value{
A tibble.
}
\description{
\code{repo_contributors} Returns a data frame containing statistics on contributor(s) to the provided repositories.
}
\examples{
\dontrun{
repo_contributors("ghclass-test/test")
}

}
