% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_create.R
\name{repo_create}
\alias{repo_create}
\title{Create repository}
\usage{
repo_create(org, name, prefix = "", suffix = "", private = TRUE,
  auto_init = FALSE, gitignore_template = "R")
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{name}{Character. One or more GitHub username or team name.}

\item{prefix}{Character. Common repository name prefix}

\item{suffix}{Character. Common repository name suffix}

\item{private}{Logical. Create private repositories?}

\item{auto_init}{Logical. Initialize the repository with a README.md?}

\item{gitignore_template}{Character. .gitignore template language.}
}
\description{
\code{repo_create} creates either individual or team repositories for a given
assignment.
}
\examples{
\dontrun{
repo_create("ghclass-test", c("user01","user02"), prefix = "hw01-")
}

}
\seealso{
Other github repository related functions: \code{\link{repo_delete}}
}
\concept{github repository related functions}
