% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_delete_file.R
\name{repo_delete_file}
\alias{repo_delete_file}
\title{Function for deleting files from a Github repository}
\usage{
repo_delete_file(repo, path, message = NULL, branch = "master")
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{path}{Character. File path within the repository.}

\item{message}{Character. Commit message. If not provided, a custom character string will be created.}

\item{branch}{Character. Name of branch to use, defaults to "master".}
}
\description{
Function for deleting files from a Github repository
}
\examples{
\dontrun{
repo = repo_create("ghclass-demo", "hw1-demo")
repo_mirror("Sta323-Sp19/hw1", repo)

repo_delete_file("ghclass-demo/hw1-demo", "README.md")
repo_delete_file("ghclass-demo/hw1-demo", "README.md") # Fails since the file does not exist

repo_delete(repo, prompt=FALSE)
}

}
