% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_exists.R
\name{repo_exists}
\alias{repo_exists}
\title{Check existence of GitHub repository}
\usage{
repo_exists(repo, strict = FALSE, verbose = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{strict}{Logical. Specifies whether renamed repositories are allowed.}

\item{verbose}{Logical. Specifies if details on renamed repositories should be printed.}
}
\value{
A logical vector
}
\description{
\code{repo_exists} returns TRUE if the github repository exists.
The function also prints a message if a repository has been renamed.
}
\examples{
\dontrun{
repo_exists(c("rundel/ghclass", "rundel/ghclass_fake"))
}

}
