% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{repo_file_mirror}
\alias{repo_file_mirror}
\title{Mirror file(s) between repos}
\usage{
repo_file_mirror(source_repo, target_repo, path = NULL,
  source_folder = NULL, target_folder = NULL, message = NULL,
  branch = "master", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{target_repo}{Character. Address of repository in "owner/name" format.}

\item{path}{Character or character vector. Name(s) of file(s) to be moved.}

\item{source_folder}{Character. Name of folder containing file on \code{source_repo}.}

\item{target_folder}{Character. Name of folder containing file on \code{target_repo}.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch to use, defaults to "master".}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to FALSE.}

\item{verbose}{Logical. Should success / failure messages be printed.}
}
\description{
\code{repo_file_mirror} mirrors select file(s) between repositories.
}
\seealso{
Other peer review functions: \code{\link{peer_assign}},
  \code{\link{peer_file_add_aut}},
  \code{\link{peer_file_add_rev}},
  \code{\link{peer_form_create_rating}},
  \code{\link{peer_form_create_review}},
  \code{\link{peer_init}}, \code{\link{peer_return}},
  \code{\link{peer_roster_create}},
  \code{\link{peer_score_rating}},
  \code{\link{peer_score_review}}
}
\concept{peer review functions}
