% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_get_file.R
\name{repo_get_file}
\alias{repo_get_file}
\alias{get_file}
\title{Low level function for retrieving a file from a GitHub Repository}
\usage{
repo_get_file(repo, path, branch = "master")
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{branch}{Character. Name of branch to use, defaults to "master".}

\item{file}{Characer. Path to the file within the repository.}
}
\description{
Low level function for retrieving a file from a GitHub Repository
}
