% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_ignore.R
\name{repo_ignore}
\alias{repo_ignore}
\title{Ignore repository}
\usage{
repo_ignore(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
Ignores the provided GitHub repositories.
}
