% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_get_template.R
\name{repo_is_template}
\alias{repo_is_template}
\title{Check if a GitHub repository is a template repository}
\usage{
repo_is_template(repo)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}
}
\value{
A logical vector
}
\description{
\code{repo_is_template} returns TRUE if the github repository is a template repository.
}
\examples{
\dontrun{
repo_is_template(c("rundel/ghclass", "rundel/ghclass_fake"))
}

}
