% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_mirror.R
\name{repo_mirror}
\alias{repo_mirror}
\alias{mirror_repo}
\title{Mirror repository}
\usage{
repo_mirror(source_repo, target_repo, verbose = FALSE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{target_repo}{Character. One or more repository addresses in "owner/name" format.}

\item{verbose}{Logical. Display verbose output.}
}
\description{
\code{repo_mirror} mirrors the content of one repository to another repository, or set of
repositories.
}
\examples{
\dontrun{
repo_mirror("ghclass-test/hw1", c("ghclass-test/hw1-Team1", "ghclass-test/hw1-Team2"))
repo_mirror("ghclass-test/hw1", org_repos("ghclass-test","hw1-"))
}

}
