% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{repo_mirror_file}
\alias{repo_mirror_file}
\title{Mirror file(s) between repos}
\usage{
repo_mirror_file(source_repo, target_repo, path = NULL,
  source_folder = NULL, target_folder = NULL, message = NULL,
  branch = "master", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{target_repo}{Character. Address of repository in "owner/name" format.}

\item{path}{Character or character vector. Name(s) of file(s) to be moved.}

\item{source_folder}{Character. Name of folder containing file on \code{source_repo}.}

\item{target_folder}{Character. Name of folder containing file on \code{target_repo}.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch to use, defaults to "master".}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to FALSE.}
}
\description{
\code{repo_mirror_file} mirrors select file(s) between repositories.
}
