% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_mirror_template.R
\name{repo_mirror_template}
\alias{repo_mirror_template}
\title{Mirror template repository}
\usage{
repo_mirror_template(template_repo, repo, private = TRUE)
}
\arguments{
\item{template_repo}{Character. Address of template repository in \code{owner/name} format.}

\item{repo}{Character. One or more repository addresses in \code{owner/name} format.
Note when using template repos these new repositories must not already exist.}

\item{private}{Logical. Should the new repository be private or public.}
}
\description{
\code{repo_mirror_template} mirrors the content of a source (template) repository to a new (not yet created)
repository, or vector of repositories.
}
