% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_n_commits.R
\name{repo_n_commits}
\alias{repo_n_commits}
\title{Get repository commit count}
\usage{
repo_n_commits(repo, branch = "master", quiet = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{branch}{Character.    Branch to list commits from.}

\item{quiet}{Logical. Should an error message be printed if a repo does not exist.}
}
\description{
\code{repo_n_commits} returns a tibble of repositories and the number of commits in a given branch.
}
\examples{
\dontrun{
repo_n_commits("rundel/ghclass")
}

}
