% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_prs.R
\name{repo_prs}
\alias{repo_prs}
\title{Get repository pull requests}
\usage{
repo_prs(repo, state = c("open", "closed", "all"))
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}
}
\description{
\code{repo_prs} returns a data fram of repository pull requests.
}
