% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_badge.R
\name{repo_remove_badge}
\alias{repo_remove_badge}
\title{Remove one or more GitHub Action badges}
\usage{
repo_remove_badge(repo, workflow = ".*?")
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{workflow}{Character. Name of the workflow badge to be removed, or a regex pattern that matches the workflow name.}
}
\description{
Remove one or more GitHub Action badges
}
