% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_remove_user.R
\name{repo_remove_user}
\alias{repo_remove_user}
\title{Remove a user from a repository}
\usage{
repo_remove_user(repo, user)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{user}{Character. One or more GitHub usernames.}
}
\description{
Remove a user from a repository
}
