% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_rename.R
\name{repo_rename}
\alias{repo_rename}
\alias{rename_repo}
\title{Rename repository}
\usage{
repo_rename(repo, new_name)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{new_name}{Character. New name of repository in the "name" format.}
}
\description{
\code{repo_rename} renames repositories. Use with caution as repositories retain their
unique identifier upon renaming and can be accessed under their old names due to
GitHub re-directing.
}
\examples{
\dontrun{
repo_rename("ghclass-test/hw1", "homework1")
}

}
