% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_set_template.R
\name{repo_set_template}
\alias{repo_set_template}
\title{Change the template status of a GitHub repository}
\usage{
repo_set_template(repo, status = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{status}{Logical. Should the repository be set as a template repository}
}
\description{
\code{repo_set_template} returns TRUE if the github repository is a template repository.
}
