% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_watch.R
\name{repo_watch}
\alias{repo_watch}
\title{Watch repository}
\usage{
repo_watch(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
Watches / subscribes to the provided GitHub repositories.
}
\seealso{
Other notification: \code{\link{repo_ignore}},
  \code{\link{repo_unwatch}}, \code{\link{repo_watching}}
}
\concept{notification}
