% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_watching.R
\name{repo_watching}
\alias{repo_watching}
\alias{get_watching}
\title{Get watched repos}
\usage{
repo_watching(filter = NULL, exclude = FALSE)
}
\arguments{
\item{filter}{character, regex pattern for matching (or excluding) repositories.}

\item{exclude}{logical, should entries matching the regex be excluded or included.}
}
\description{
Returns all of the authenticated user's watched repositories. This should
match the list at \href{https://github.com/watching}{github.com/watching}.
The function can also filter the results for matching (or excluding)
repositories.
}
\examples{
\dontrun{
repo_watching()
repo_watching("hw1")
}

}
