% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_reset_token.R
\name{reset_github_token}
\alias{reset_github_token}
\title{Reset github token}
\usage{
reset_github_token()
}
\description{
\code{resset_github_token} removes any value stored in the \code{GITHUB_PAT}
enivronmental variable.
}
\examples{
\dontrun{
token = "0123456789ABCDEF0123456789ABCDEF01234567"
set_github_token("0123456789ABCDEF0123456789ABCDEF01234567")
get_github_token() == token
reset_github_token()
get_github_token() == token
}

}
\seealso{
Other authentication: \code{\link{get_github_token}},
  \code{\link{set_github_token}},
  \code{\link{test_github_token}}, \code{\link{with_pat}}
}
\concept{authentication}
