% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_set_token.R
\name{set_github_token}
\alias{set_github_token}
\title{Set github token}
\usage{
set_github_token(token)
}
\arguments{
\item{token}{character, either the path of a file contained the token or the actual token.}
}
\description{
\code{set_github_token} defines the user's github authentication token by
defining the \code{GITHUB_PAT} enivronmental variable. This value can then
be subsequently accessed using \code{get_github_token}.
}
\examples{
\dontrun{
set_github_token("~/.github/token")
set_github_token("0123456789ABCDEF0123456789ABCDEF01234567")
}

}
\seealso{
Other authentication: \code{\link{get_github_token}},
  \code{\link{github_reset_token}},
  \code{\link{test_github_token}}, \code{\link{with_pat}}
}
\concept{authentication}
