% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{set_wercker_token}
\alias{set_wercker_token}
\title{Set wercker token}
\usage{
set_wercker_token(token)
}
\arguments{
\item{token}{character, either the path of a file contained the token or the actual token.}
}
\description{
\code{set_wercker_token} defines the user's wercker authentication token,
this value is then accessed usin \code{get_wercker_token}
}
\examples{
\dontrun{
set_wercker_token("~/.wercker/token")
set_wercker_token("0123456789ABCDEF0123456789ABCDEF01234567")
}

}
