% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_safely.R
\name{status_msg}
\alias{status_msg}
\title{Internal safely-result helper}
\usage{
status_msg(
  x,
  success = NULL,
  fail = NULL,
  include_error_msg = TRUE,
  .envir = parent.frame()
)
}
\description{
Internal safely-result helper
}
\keyword{internal}
