% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{style_repo}
\alias{style_repo}
\title{Style repository}
\usage{
style_repo(repo, files = c("*.R", "*.Rmd"), branch = "styler",
  base = "master", create_pull_request = TRUE,
  tag_collaborators = TRUE, git = require_git(), verbose = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{files}{Character or vector of characters. Names of .R and/or .Rmd files that styler should be applied to.}

\item{branch}{Character. Name of new branch to be created. Default is "styler".}

\item{base}{Character. Name of branch that contains the .R and/or .Rmd files to be reformatted.}

\item{create_pull_request}{Logical. If TRUE, a pull request is created from branch to base.}

\item{tag_collaborators}{Logical. If TRUE, a message with the repository collaborators is displayed.}

\item{verbose}{Logical. Display verbose output.}
}
\description{
\code{style_repo} implements "non-invasive pretty-printing of R source code" of .R or .Rmd files within a repository using the \code{styler} package and adhering to \code{tidyverse} formatting guidelines.
}
\examples{
\dontrun{
style_repo("Sta523-Fa17/base_hw1", files = c("hw1_sample.Rmd"))
}

}
