% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_create.R
\name{team_create}
\alias{team_create}
\alias{create_team}
\title{Create team(s)}
\usage{
team_create(
  org,
  team,
  prefix = "",
  suffix = "",
  privacy = c("secret", "closed")
)
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{team}{character, listing one or more teams}

\item{prefix}{Character. Common team name prefix}

\item{suffix}{Character. Common team name suffix}

\item{privacy}{character, level of privacy of teams, closed (visible to all
members of the organization) or secret (only visible to organization owners
and members of a team), default is closed}
}
\description{
\code{team_create} creates teams in your GitHub organization
}
\examples{
\dontrun{
team_create("ghclass",c("team01","team01"))
}

}
