% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_invite.R
\name{team_invite}
\alias{team_invite}
\title{Add Members to an Organizaton's Team(s)}
\usage{
team_invite(org, user, team, create_missing_teams = TRUE)
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{user}{character, one or more usernames to invite}

\item{team}{character, one or more team names}

\item{create_missing_teams}{logical, if a team does not already exist
should it be added to the GitHub organization.}
}
\description{
\code{team_invite} add members to GitHub Organization Teams.
}
\examples{
\dontrun{
team_invite("ghclass-test", "rundel", c("hw1-team01","hw1-team02"))
}

}
\seealso{
Other github organization team related functions: \code{\link{org_teams}},
  \code{\link{team_create}}, \code{\link{team_delete}},
  \code{\link{team_members}},
  \code{\link{team_pending_members}},
  \code{\link{team_rename}}
}
\concept{github organization team related functions}
