% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_pending_members.R
\name{team_pending_members}
\alias{team_pending_members}
\title{Get pending team members}
\usage{
team_pending_members(org, team = org_teams(org))
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{team}{character or data frame, listing one or more team}
}
\description{
\code{team_pending_members} returns a data frame of pending team members.
}
\examples{
\dontrun{
team_pending_members("ghclass",c("team01","team02"))
}

}
\seealso{
Other github organization team related functions: \code{\link{org_teams}},
  \code{\link{team_create}}, \code{\link{team_delete}},
  \code{\link{team_invite}}, \code{\link{team_members}},
  \code{\link{team_rename}}
}
\concept{github organization team related functions}
