% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rename.R
\name{team_rename}
\alias{team_rename}
\title{Rename existing team(s)}
\usage{
team_rename(org, team, new_team)
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{team}{character, one or more existing team names}

\item{new_team}{character, one or more new team names}
}
\description{
\code{team_rename} renames an existing team within the given GitHub organization.
}
\examples{
\dontrun{
team_rename("ghclass-test", "hw1-team01", "hw01-team01")
}

}
