% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{unprotect_branch}
\alias{unprotect_branch}
\title{Unprotect branch}
\usage{
unprotect_branch(repo, branch = "master", verbose = TRUE)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}

\item{cur_branch}{name of existing branch}

\item{new_branch}{name of branch to create}
}
\description{
\code{unprotect_branch} removes protections from the specified branch. See
\link{https://help.github.com/en/articles/about-protected-branches} for more details
on what this changes.
}
\seealso{
Other branch functions: \code{\link{create_branch}},
  \code{\link{protect_branch}}
}
\concept{branch functions}
