% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo.R
\name{update_repos}
\alias{update_repos}
\title{Update local repos via git pull}
\usage{
update_repos(repo_dirs, git = require_git(), options = "",
  verbose = FALSE)
}
\arguments{
\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your \code{PATH}, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}

\item{repos_dir}{Character vector of repo directories or a single directory containing the repo directories.}
}
\description{
\code{update_repos} uses the git binary to update the provided repositories
using git pull.
}
\examples{
\dontrun{
update_repos("hw1/")
update_repos(c("hw1/team01","hw1/team02","hw1/team03"))
}

}
\seealso{
Other local repo functions: \code{\link{clone_repos}}
}
