% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_repos.R
\name{user_repos}
\alias{user_repos}
\title{Get user's repository}
\usage{
user_repos(user, type = c("owner", "all", "public", "private", "member"),
  filter = NULL, exclude = FALSE, full_repo = TRUE)
}
\arguments{
\item{user}{Character. Username of the GitHub user.}

\item{type}{Characer. Can be one of "all", "owner", "public", "private", "member".}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression in \code{filter} be excluded or included?}

\item{full_repo}{Logical. Should the full repository address be returned (e.g. \code{owner/name} instead of just \code{repo})?}
}
\description{
\code{user_repos} returns a (filtered) vector of repositories belonging to a GitHub user.
}
\examples{
\dontrun{
user_repos("rundel")
user_repos("rundel", "ghclass")
}

}
