% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_type.R
\name{user_type}
\alias{user_type}
\title{Check type of a username}
\usage{
user_type(user)
}
\arguments{
\item{user}{Character. Username(s) to be checked.}
}
\description{
\code{user_type} returns a Character vector of the accounts' types.
}
\examples{
\dontrun{
user_type(c("rundel","ghclass-demo"))
}

}
