% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watching.R
\name{watch_repo}
\alias{watch_repo}
\title{Watch repository}
\usage{
watch_repo(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
Watches / subscribes to the provided GitHub repositories.
}
\seealso{
Other notification functions: \code{\link{get_watching}},
  \code{\link{ignore_repo}}, \code{\link{unwatch_repo}}
}
\concept{notification functions}
