% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_with_pat.R
\name{with_pat}
\alias{with_pat}
\alias{local_pat}
\title{Use Temporary GitHub Access Token}
\usage{
with_pat(new, code)

local_pat(new, .local_envir = parent.frame())
}
\arguments{
\item{new}{Temporary GitHub access token}

\item{code}{Code to execute with the temporary token}

\item{.local_envir}{The environment to use for scoping.}
}
\description{
Temporarily change the \code{GITHUB_PAT} environmental variable for
GitHub authentication. Based on the \code{withr} interface.
}
\details{
if \code{NA} is used the \code{GITHUB_PAT} environment variable will be unset.
}
\seealso{
Other authentication: \code{\link{github_get_token}},
  \code{\link{github_reset_token}},
  \code{\link{github_set_token}},
  \code{\link{github_test_token}}
}
\concept{authentication}
