% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R, R/action_add_badge.R,
%   R/action_remove_badge.R, R/action_workflows.R
\name{action}
\alias{action}
\alias{action_add_badge}
\alias{action_remove_badge}
\alias{action_workflows}
\title{GitHub Actions tools}
\usage{
action_add_badge(
  repo,
  workflow = NULL,
  where = "^.",
  line_padding = "\\n\\n\\n",
  file = "README.md"
)

action_remove_badge(repo, workflow_pat = ".*?", file = "README.md")

action_workflows(repo, full = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{workflow}{Character. Name of the workflow located in \verb{.github/workflows/}.}

\item{where}{Character. Regex pattern of where to insert the badge, defaults
to the beginning of the README}

\item{line_padding}{Character. What should be added to the end of the link,
defaults to sufficient new lines to add a blank line after the badge.}

\item{file}{Character. Target file to be modified, defaults to \code{README.md}.#'}

\item{workflow_pat}{Character. Name of the workflow badge to be removed, or a regex pattern
that matches the workflow name.}

\item{full}{Logical. Should all workflow columns be returned}
}
\description{
\itemize{
\item \code{action_workflows} - returns a data frame containing details on a repositories workflows.
\item \code{action_add_badge} - Add a GitHub Actions badge to a file.
\item \code{action_remove_badge} - Remove one or more GitHub Action badges from a file.
}
}
\examples{
\dontrun{
action_workflows("rundel/ghclass")
}

}
