% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R, R/action_artifacts.R,
%   R/action_artifacts_download.R, R/action_runs.R, R/action_status.R,
%   R/action_workflows.R
\name{action}
\alias{action}
\alias{action_artifacts}
\alias{action_artifact_download}
\alias{action_runs}
\alias{action_status}
\alias{action_workflows}
\title{Retrieve information about GitHub Actions workflows and their runs.}
\usage{
action_artifacts(repo, keep_expired = FALSE, which = c("latest", "all"))

action_artifact_download(
  repo,
  dir,
  ids = action_artifacts(repo),
  keep_zip = FALSE,
  file_pat = "",
  overwrite = FALSE
)

action_runs(repo, limit = 50)

action_status(repo, branch = NULL, before = NULL, limit = 50)

action_workflows(repo, full = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{keep_expired}{Logical. Should expired artifacts be returned.}

\item{which}{Character. Either \code{"latest"} to return only the most recent of each
artifact or \code{"all"} to return all artifacts.}

\item{dir}{Character. Path to the directory where artifacts will be saved.}

\item{ids}{Integer or data frame. Artifact ids to be downloaded from the repositories.
If a data frame is passed then the \code{id} column will be used.}

\item{keep_zip}{Logical. Should the artifact zips be saved (\code{TRUE}) or their contents (\code{FALSE}).}

\item{file_pat}{Character. If extracting zip with multiple files, regexp pattern to match filename.}

\item{overwrite}{Logical. Should existing files be overwritten.}

\item{limit}{Numeric. Maximum number of workflow runs to return.}

\item{branch}{Character. Vector of branches to report. Default \code{NULL}, returns all branches.}

\item{before}{Datetime. Remove all runs created after this timestamp.
Character values are converted using \code{lubridate::ymd_hms()}.}

\item{full}{Logical. Should all workflow columns be returned. Default \code{FALSE}.}
}
\value{
\code{action_workflows()}, \code{action_runs()}, \code{action_status()}, and \code{action_artifacts}
all return tibbles containing information on requested repos' available workflows,
recent action runs, recent action run statuses respectively, and generated artifacts
respectively.

\code{action_artifact_download()} returns a character vector containing the paths of all
downloaded fules
}
\description{
\itemize{
\item \code{action_workflows()} - retrieve details on repo workflows.
\item \code{action_runs()} - retrieve details on repo workflow runs.
\item \code{action_status()} - retrieve details on most recent workflow runs.
\item \code{action_artifacts()} - retrieve details on available workflow artifacts.
\item \code{action_artifact_download()} - downloads artifact(s) into a local directory.
}
}
\examples{
\dontrun{
action_workflows("rundel/ghclass")

action_runs("rundel/ghclass")

action_status(c("rundel/ghclass", "rundel/parsermd"))

action_artifacts(c("rundel/ghclass", "rundel/parsermd"))
}

}
