% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_workflows.R
\name{action}
\alias{action}
\alias{action_workflows}
\title{Return a data frame containing details on a repository workflows.}
\usage{
action_workflows(repo, full = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{full}{Logical. Should all workflow columns be returned}
}
\description{
Return a data frame containing details on a repository workflows.
}
\examples{
\dontrun{
action_workflows("rundel/ghclass")
}

}
