% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_add_file.R
\name{add_file}
\alias{add_file}
\title{Add files to a repo}
\usage{
add_file(repo, file, message = NULL, branch = "master",
  preserve_path = FALSE, overwrite = FALSE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{file}{Character. Local file path(s) of file or files to be added.}

\item{message}{Character. Commit message. If not provided, a custom character string will be created, in the form of "Added file(s): filename(s)". If this custom message character length exceeds 50, it will be shortened to "Added file(s)".}

\item{branch}{Character. Name of branch to use, defaults to "master".}

\item{preserve_path}{Logical. Should the local relative path be preserved.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to FALSE.}
}
\description{
\code{add_file} uses the GitHub API to add/update files in an existing repo on GitHub. Note that due to time delays in caching, files that have been added very recently might not yet be displayed as existing and might accidentally be overwritten.
}
\examples{
\dontrun{
add_file("rundel/ghclass", "Update DESCRIPTION", "./DESCRIPTION")
}

}
\seealso{
Other file functions: \code{\link{add_content}},
  \code{\link{get_file}}, \code{\link{get_readme}},
  \code{\link{put_file}}
}
\concept{file functions}
