% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R, R/branch_create.R, R/branch_delete.R
\name{branch}
\alias{branch}
\alias{branch_create}
\alias{branch_delete}
\alias{branch_remove}
\title{Repository branch tools

\code{branch_create} creates a new branch from an existing GitHub repo.

\code{branch_delete} deletes a branch from an existing GitHub repo.}
\usage{
branch_create(repo, branch = "master", new_branch)

branch_delete(repo, branch)

branch_remove(repo, branch)
}
\arguments{
\item{repo}{Github repository address in \code{owner/repo} format.}

\item{branch}{Repository branch to use.}

\item{new_branch}{Name of branch to create.}
}
\description{
Repository branch tools

\code{branch_create} creates a new branch from an existing GitHub repo.

\code{branch_delete} deletes a branch from an existing GitHub repo.
}
\examples{
\dontrun{
repo_create("ghclass-demo", "test_branch", auto_init=TRUE)

branch_create("ghclass-demo/test_branch", new_branch = "test")
repo_branches("ghclass-demo/test_branch")

branch_delete("ghclass-demo/test_branch", branch="test")
repo_branches("ghclass-demo/test_branch")

repo_delete("ghclass-demo/test_branch", prompt = FALSE)
}

}
\seealso{
\link{repo_branches}
}
