% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_protect.R, R/branch_protect_review.R
\name{branch_unprotect}
\alias{branch_unprotect}
\alias{branch_protect_with_review}
\title{Branch protection}
\usage{
branch_unprotect(repo, branch)

branch_protect_with_review(repo, branch)
}
\arguments{
\item{repo}{Github repository address in \code{owner/repo} format.}

\item{branch}{Repository branch to use.}
}
\value{

}
\description{
Adds or removes branch protection on the repo. Specifically,
added branch protection enforces 1 approving
PR review from Codeowners before allowing a merge.
This is typically coupled with a \code{.github/CODEOWNERS}
file in the repo root (on \code{main/master}).
}
\section{Functions}{
\itemize{
\item \code{branch_unprotect}: Removes any branch protection
}}

\examples{
\dontrun{
repo_create("ghclass-test", "protect_me", auto_init=TRUE)

repo_put_file("ghclass-test/protect_me", ".github/CODEOWNERS",
  content = "* @dajmcdon")
branch_protect_with_review("ghclass-test/protect_me", "main")
}
}
