% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_files.R
\name{check_file_modification}
\alias{check_file_modification}
\title{Check for modifications of file on GitHub}
\usage{
check_file_modification(repo, file, include_admin)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{file}{Character. Name of file.}

\item{include_admin}{Logical. Should users with admin privileges be included in checking for previous file modifications.}
}
\value{
TRUE or FALSE
}
\description{
\code{check_file_modification} checks whether a file on GitHub has previously been modified by a commit (not taking into account the initial commit).
}
