% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{check_repo}
\alias{check_repo}
\title{Check existance of github repo}
\usage{
check_repo(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
\code{check_repo} returns TRUE if the github repository exists
}
\examples{
\dontrun{
check_repo("rundel/ghclass")
check_repo("rundel/ghclass_fake")
}

}
\concept{github repo related functions}
