% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{check_repo}
\alias{check_repo}
\title{Check existance of github repo}
\usage{
check_repo(repo)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format. Can be vector or list of characters.}
}
\value{
logical
}
\description{
\code{check_repo} returns TRUE if the GitHub repository exists.
}
\examples{
\dontrun{
check_repo("rundel/ghclass")
check_repo("rundel/ghclass_fake")
check_repo(c("rundel/ghclass", "rundel/ghclass_fake"))
check_repo(list("rundel/ghclass", "rundel/ghclass_fake"))
}

}
\seealso{
Other github repo related functions: \code{\link{get_watching}},
  \code{\link{ignore_repo}}, \code{\link{unwatch_repo}},
  \code{\link{watch_repo}}
}
\concept{github repo related functions}
