% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{check_repo}
\alias{check_repo}
\title{Check existence of GitHub repository}
\usage{
check_repo(repo, strict = FALSE, verbose = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{strict}{Logical. Specifies whether renamed repositories are allowed.}

\item{verbose}{Logical. Specifies if details on renamed repositories should be printed.}
}
\value{
A logical vector
}
\description{
\code{check_repo} returns TRUE if the github repository exists.
The function also prints a message if a repository has been renamed.
}
\examples{
\dontrun{
check_repo(c("rundel/ghclass", "rundel/ghclass_fake"))
}

}
\concept{github repo related functions}
