% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo.R
\name{clone_repos}
\alias{clone_repos}
\alias{grab_repos}
\title{Clone repos from github}
\usage{
clone_repos(repos, local_path = "./", git = require_git(), options = "")
}
\arguments{
\item{repos}{Character vector of repo names with the form \emph{owner/name}.}

\item{local_path}{Local directory into which the repos will be cloned.}

\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your \code{PATH}, it will throw an error if git cannot be found.}

\item{options}{git binary options (e.g. \code{--bare})}
}
\description{
\code{clone_repos} uses the git binary to clone the provided repositories
into a local directory.
}
\examples{
\dontrun{
clone_repos("rundel/ghclass")
}

}
