% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{create_individual_repo}
\alias{create_individual_repo}
\title{Create individual repositories}
\usage{
create_individual_repo(org, user, prefix = "", suffix = "",
  private = TRUE, verbose = TRUE, auto_init = FALSE,
  gitignore_template = "R")
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{user}{character or data frame, listing one or more users}

\item{prefix}{character, resulting repo name will start with this character string}

\item{suffix}{character, resulting repo name will end with this character string}

\item{private}{logical, create private repos}

\item{verbose}{logical, display verbose output}

\item{auto_init}{logical, initialize the repository with a README.md}

\item{gitignore_template}{character, .gitignore template language}
}
\description{
\code{create_individual_repo} creates repos for each student for a given
assignment
}
\examples{
\dontrun{
create_individual_repo("ghclass",c("user01","user02"), prefix="hw01-")
}

}
\seealso{
Other github organization related functions: \code{\link{create_team_repo}},
  \code{\link{create_team}}, \code{\link{get_members}},
  \code{\link{get_pending_members}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_repos}}, \code{\link{get_team_members}},
  \code{\link{get_team_repos}}, \code{\link{get_teams}},
  \code{\link{invite_user}}
}
\concept{github organization related functions}
