% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{create_team}
\alias{create_team}
\title{Create team(s)}
\usage{
create_team(org, team = character(), privacy = c("closed", "secret"),
  verbose = TRUE)
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{team}{character, listing one or more teams}

\item{privacy}{character, level of privacy of teams, closed (visible to all
members of the organization) or secret (only visible to organization owners
and members of a team), default is closed}
}
\description{
\code{create_team} creates teams in your organization
}
\examples{
\dontrun{
create_team("ghclass",c("team01","team01"))
}

}
\seealso{
Other github organization related functions: \code{\link{create_individual_repo}},
  \code{\link{create_team_repo}},
  \code{\link{get_members}},
  \code{\link{get_pending_members}},
  \code{\link{get_pending_team_members}},
  \code{\link{get_repos}}, \code{\link{get_team_members}},
  \code{\link{get_team_repos}}, \code{\link{get_teams}},
  \code{\link{invite_user}}
}
\concept{github organization related functions}
