% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{create_team_repo}
\alias{create_team_repo}
\title{Create team repositories}
\usage{
create_team_repo(org, team, prefix = "", suffix = "", private = TRUE,
  verbose = TRUE, auto_init = FALSE, gitignore_template = "R")
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{team}{Character or data frame. Vector of team names.}

\item{prefix}{Character. Resulting repo name will start with this character string.}

\item{suffix}{Character. Resulting repo name will end with this character string.}

\item{private}{Logical. Create private repos.}

\item{verbose}{Logical. Display verbose output.}

\item{auto_init}{Logical. Initialize the repository with a README.md.}

\item{gitignore_template}{Character. .gitignore template language.}
}
\description{
\code{create_team_repo} creates repos for team(s).
}
\examples{
\dontrun{
create_team_repo("Sta523-Fa17", c("team01","team02"), prefix = "hw01-")
}

}
\seealso{
Other github organization related functions: \code{\link{create_individual_repo}},
  \code{\link{get_members-deprecated}},
  \code{\link{get_member}},
  \code{\link{get_pending_members-deprecated}},
  \code{\link{get_pending_member}},
  \code{\link{get_repos-deprecated}},
  \code{\link{get_repo}}, \code{\link{get_team_repos}},
  \code{\link{invite_user}}
}
\concept{github organization related functions}
