% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_admin}
\alias{get_admin}
\title{List repository administrators}
\usage{
get_admin(org, verbose = FALSE)
}
\arguments{
\item{org}{Character. Name of GitHub organization.}

\item{verbose}{Logical. Display verbose output.}
}
\value{
A list containing a character vector of repository administrators.
}
\description{
\code{get_admin} creates a list of repository administrators.
}
\examples{
\dontrun{
get_admin("Sta523-Fa17")
}

}
