% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{get_collaborator}
\alias{get_collaborator}
\title{List collaborator(s)}
\usage{
get_collaborator(repo, include_admin = TRUE)
}
\arguments{
\item{repo}{Character. Address of one or more repositories in \code{owner/name} format.}

\item{include_admin}{Logical. If \code{FALSE}, user names of users with Admin rights are not included, defaults to \code{TRUE}.}
}
\value{
A tibble with two columns, \code{repo} and \code{username}.
}
\description{
\code{get_collaborator} Returns a vector of collaborator user names. Users with admin rights are by default excluded.
}
\examples{
\dontrun{
get_collaborators("ghclass-test/test2")
}

}
