% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{get_pending_team_members}
\alias{get_pending_team_members}
\title{Get pending team members}
\usage{
get_pending_team_members(org, teams = get_teams(org))
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{teams}{character or data frame, listing one or more team}
}
\description{
\code{get_pending_team_members} returns a data frame of pending team members.
}
\examples{
\dontrun{
get_pending_team_members("ghclass",c("team01","team02"))
}

}
\seealso{
Other github organization related functions: \code{\link{get_members}},
  \code{\link{get_pending_members}},
  \code{\link{get_repos}}, \code{\link{get_team_members}},
  \code{\link{get_team_repos}}, \code{\link{get_teams}}
}
