% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghclass.R
\docType{package}
\name{ghclass}
\alias{ghclass}
\title{ghclass: A package for github based classroom and assignment management}
\description{
This package is for everyone! But really, if you’re an instructor who
uses GitHub for your class management, e.g. students submit assignments
via GitHub repos, this package is definitely for you! The package also
assumes that you’re an R user, and you probably teach R as well, though
that’s not a requirement since this package is all about setting up
repositories with the right permissions, not what your students put in
those repositories.
}
