% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghclass.R
\docType{package}
\name{ghclass}
\alias{ghclass}
\alias{ghclass-package}
\title{ghclass: A package for github based classroom and assignment management}
\description{
ghclass: A package for github based classroom and assignment management
}
\section{github repo functions}{


\itemize{
\item \code{One} -
\item \code{Two} -
}
}

\section{local repo functions}{


\itemize{
\item \code{\link{clone_repos}} - clone repos locally
\item \code{\link{update_repos}} - update local repos via pull
}
}

\section{github organization functions}{


\itemize{
\item \code{One} -
\item \code{Two} -
}
}

\section{wercker functions}{


\itemize{
\item \code{One} -
\item \code{Two} -
}
}

\section{authentication functions}{


GitHub
\itemize{
\item \code{\link{get_github_token}} - get github token
\item \code{\link{set_github_token}} - set github token
\item \code{\link{test_github_token}} - test github token
}
Wercker
\itemize{
\item \code{\link{get_wercker_token}} - get wercker token
\item \code{\link{set_wercker_token}} - set wercker token
\item \code{\link{test_wercker_token}} - test wercker token
}
}

