% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api_limit.R
\name{github_api_limit}
\alias{github_api_limit}
\alias{github_get_api_limit}
\alias{github_set_api_limit}
\title{Tools for limiting gh's GitHub api requests.}
\usage{
github_get_api_limit()

github_set_api_limit(limit = 1000L)
}
\arguments{
\item{limit}{The maximum number of records to return from an API request.}
}
\description{
\itemize{
\item \code{github_get_api_limit} - returns the current limit on results returned by gh.
\item \code{github_set_api_limit} - sets a limit on results returned by gh.
}
}
\details{
This value is stored in the "ghclass.api.limit" option globally.
}
\examples{
github_get_api_limit()

github_set_api_limit(500)

github_get_api_limit()
}
