% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_whoami.R
\name{github_orgs}
\alias{github_orgs}
\title{Collect details on the authenticated user's GitHub organization memberships
(based on the current PAT).}
\usage{
github_orgs(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Should status messages be shown.}
}
\value{
Returns a tibble with organization details.
}
\description{
Collect details on the authenticated user's GitHub organization memberships
(based on the current PAT).
}
\examples{
\dontrun{
github_orgs()
}

}
