% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_test_token.R
\name{github_test_token}
\alias{github_test_token}
\title{Test github token}
\usage{
github_test_token(token = get_github_token())
}
\arguments{
\item{token}{github api personal access token}
}
\description{
github_test_token` checks if a token is valid by attempting to authenticate
with the GitHub api.
}
\examples{
\dontrun{
github_test_token()
github_test_token("bad_token")
}

}
\seealso{
Other authentication: \code{\link{get_github_token}},
  \code{\link{github_reset_token}},
  \code{\link{github_set_token}}, \code{\link{with_pat}}
}
\concept{authentication}
