% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_token.R
\name{github_token}
\alias{github_token}
\alias{github_get_token}
\alias{github_set_token}
\alias{github_reset_token}
\alias{github_test_token}
\title{Tools for handling GitHub personal access tokens (PAT)}
\usage{
github_get_token()

github_set_token(token)

github_reset_token()

github_test_token(token = github_get_token())
}
\arguments{
\item{token}{Character. Either the literal token, or the path to a file containing the token.}
}
\description{
\itemize{
\item \code{github_get_token} - returns the user's GitHub personal access token (PAT).
\item \code{github_set_token} - defines the user's GitHub PAT by setting the \code{GITHUB_PAT} environmental variable.
This value will persist until the session ends or \code{gihub_reset_token()} is called.
\item \code{github_reset_token} - removes the value stored in the \code{GITHUB_PAT} environmental variable.
\item \code{github_test_token} - checks if a PAT is valid by attempting to authenticate with the GitHub API
}
}
\details{
This package looks for the personal access token (PAT) in the following places (in order):
\itemize{
\item Value of \code{GITHUB_PAT} environmental variable.
\item Any GitHub PAT token(s) stored with \code{gitcreds} via \code{gitcreds_set()}.
}

For additional details on creating a GitHub PAT see the usethis vignette on
\href{https://usethis.r-lib.org/articles/articles/git-credentials.html}{Managing Git(Hub) Credentials}.
For those who do not wish to read the entire article, the quick start method is to use:
\itemize{
\item \code{usethis::create_github_token()} - to create the token and then,
\item \code{gitcreds::gitcreds_set()} - to securely cache the token.
}
}
\examples{
\dontrun{
github_test_token()

(pat = github_get_token())

github_set_token("ghp_BadTokenBadTokenBadTokenBadTokenBadToken")
github_get_token()
github_test_token()

github_set_token(pat)
}

}
