% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_token.R
\name{github_token}
\alias{github_token}
\alias{github_get_token}
\alias{github_set_token}
\alias{github_reset_token}
\alias{github_test_token}
\title{Tools for handling GitHub personal access tokens (PAT)}
\usage{
github_get_token()

github_set_token(token)

github_reset_token()

github_test_token(token = github_get_token())
}
\arguments{
\item{token}{Character. Either the literal token path, or the path to a file containing the token.}
}
\description{
\itemize{
\item \code{github_get_token} - returns the user's GitHub personal access token (PAT).
\item \code{github_set_token} - defines the user's GitHub PAT by setting the \code{GITHUB_PAT} enivronmental variable. This value can then
\item \code{github_reset_token} - removes the value stored in the \code{GITHUB_PAT} environmental variable.
\item \code{github_test_token} - checks if a PAT is valid by attempting to authenticate with the GitHub API
}
}
\details{
This package looks for the personal access token (PAT) in the following places (in order):
\itemize{
\item Value of \code{GITHUB_PAT} environmental variable.
\item Value of \code{GITHUB_TOKEN} environmental variable.
\item Contents of \verb{~/.github/token} file.
}
}
\examples{
# Set and get the PAT
github_set_token("~/.github/token")
github_set_token("0123456789ABCDEF")
pat = github_get_token()
pat

# Test the PAT
github_reset_token()
try(github_test_token())
try(github_test_token("bad_token"))
}
