% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_whoami.R
\name{github_whoami}
\alias{github_whoami}
\title{Returns the login of the authenticated user (based on the current PAT).}
\usage{
github_whoami(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Should status messages be shown.}
}
\description{
Returns the login of the authenticated user (based on the current PAT).
}
\examples{
\dontrun{
github_whoami()
}

}
