% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watching.R
\name{ignore_repo}
\alias{ignore_repo}
\title{Ignore repository}
\usage{
ignore_repo(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
Ignores the provided GitHub repositories.
}
\seealso{
Other github repo related functions: \code{\link{check_repo}},
  \code{\link{get_watching}}, \code{\link{unwatch_repo}},
  \code{\link{watch_repo}}
}
\concept{github repo related functions}
