% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{invite_user}
\alias{invite_user}
\title{Invite user(s)}
\usage{
invite_user(org, user)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{user}{Character, character vector, or list. Listing one or more user names.}
}
\description{
\code{invite_user} invites user(s) to your organization.
}
\examples{
\dontrun{
user = c("Alice","Bob","Carol","Dave","Eve")
invite_user("Sta523-Fa17", user)
}

}
\seealso{
Other github organization related functions: \code{\link{create_repo}},
  \code{\link{get_member}},
  \code{\link{get_pending_member}}, \code{\link{get_repo}},
  \code{\link{get_team_repos}}
}
\concept{github organization related functions}
