% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{invite_user}
\alias{invite_user}
\title{Invite user(s)}
\usage{
invite_user(org, user)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{user}{character or data frame, listing one or more users}
}
\description{
\code{invite_user} invites users to your organization
}
\examples{
\dontrun{
users = c("Alice","Bob","Carol","Dave","Eve")
invite_user("ghclass", users)
}

}
\seealso{
Other github organization related functions: \code{\link{create_individual_repo}},
  \code{\link{create_team_repo}},
  \code{\link{get_members}},
  \code{\link{get_pending_members}},
  \code{\link{get_repos}}, \code{\link{get_team_repos}}
}
\concept{github organization related functions}
