% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue.R, R/issue_close.R, R/issue_create.R
\name{issue}
\alias{issue}
\alias{issue_close}
\alias{issue_create}
\title{GitHub Issue related tools}
\usage{
issue_close(repo, number)

issue_create(repo, title, body, labels = character(), assignees = character())
}
\arguments{
\item{repo}{Character. Address of one or more repositories in \code{owner/name} format.}

\item{number}{Integer. GitHub issue number of the issue.}

\item{title}{Character. Title of the issue.}

\item{body}{Character. Content of the issue.}

\item{labels}{Character. Vector of the labels to associate with this issue}

\item{assignees}{Character. Vector of logins for users assigned to the issue.}
}
\description{
\itemize{
\item \code{issue_create} creates an issue for a GitHub repository.
\item \code{issue_close} closes an issue for a GitHub repository.
}
}
\examples{
\dontrun{
repo_create("ghclass-test","test_issue")

issue_create(
  "ghclass-test/test_issue",
  title = "Issue 1",
  body = "This is an issue"
 )

issue_create(
  "ghclass-test/test_issue",
  title = "Issue 2", body = "This is also issue",
  label = "Important"
)

issue_create(
  "ghclass-test/test_issue",
  title = "Issue 3", body = "This is also issue",
  label = c("Important", "Super Important"),
  assignees = "rundel"
)

issue_close("ghclass-test/test_issue", 1)

repo_delete("ghclass-test/test_issue", prompt=FALSE)
}
}
\seealso{
\link{repo_issues}
}
