% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo.R
\name{local_repo}
\alias{local_repo}
\alias{clone_repos}
\alias{add_repos}
\alias{commit_repos}
\alias{push_repos}
\alias{pull_repos}
\title{Local repository tools}
\arguments{
\item{repos}{GitHub repo names with the form \emph{owner/name}.}

\item{repo_dirs}{Vector of repo directories or a single directory containing one or more repos.}

\item{message}{commit message}

\item{git}{Path to the local git binary. \code{require_git()} attempts to
find the git binary based on your \code{PATH}, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}
}
\description{
The functions provide tools for working with local git repositories, ghclass includes support for following git commands:
\itemize{
\item git clone = \code{clone_repos}
\item git add = \code{add_repos}
\item git commit = \code{commit_repos}
\item git push = \code{push_repos}
\item git pull = \code{pull_repos}
}
}
\examples{
\dontrun{
g = get_repos("Sta323-Sp18","hw3-")
clone_repos(g, "hw3")

pull_repos(g, "hw3")

# After Modifying hw3.Rmd
add_repos("hw3", "hw3.Rmd")
commit_repos("hw3", "Revised homework")
push_repos("hw3")
}

}
