% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo_add.R, R/local_repo_clone.R,
%   R/local_repo_commit.R, R/local_repo_docs.R, R/local_repo_pull.R,
%   R/local_repo_push.R
\name{add_repo}
\alias{add_repo}
\alias{clone_repo}
\alias{commit_repo}
\alias{local_repo}
\alias{local_repo_rename}
\alias{pull_repo}
\alias{push_repo}
\title{Local repository tools}
\usage{
add_repo(repo_dir, files = ".", git = require_git(),
  options = character(), verbose = FALSE)

clone_repo(repo, local_path = "./", branch = "master",
  git = require_git(), options = character(), absolute_path = TRUE,
  verbose = FALSE)

commit_repo(repo_dir, message, git = require_git(),
  options = character(), verbose = FALSE)

pull_repo(repo_dir, remote = "origin", branch = "master",
  git = require_git(), options = character(), verbose = FALSE)

push_repo(repo_dir, remote = "origin", branch = "master",
  git = require_git(), options = character(), verbose = FALSE)
}
\arguments{
\item{repo_dir}{Vector of repo directories or a single directory containing one or more repos.}

\item{git}{Path to the local git binary. \code{\link[=require_git]{require_git()}} attempts to
find the git binary based on your `PATH``, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}

\item{message}{commit message}

\item{repos}{GitHub repo names with the form \code{owner/name}.}
}
\description{
The functions provide tools for working with local git repositories, ghclass includes support for following git commands:
\itemize{
\item git clone = \code{clone_repo}
\item git add = \code{add_repo}
\item git commit = \code{commit_repo}
\item git push = \code{push_repo}
\item git pull = \code{pull_repo}
}
}
\examples{
\dontrun{
g = org_repos("Sta323-Sp18","hw3-")
clone_repo(g, "hw3")

pull_repo(g, "hw3")

# After Modifying hw3.Rmd
add_repo("hw3", "hw3.Rmd")
commit_repo("hw3", "Revised homework")
push_repo("hw3")
}

}
