% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_local.R
\name{local_repo}
\alias{local_repo}
\alias{clone_repo}
\alias{add_repo}
\alias{commit_repo}
\alias{push_repo}
\alias{pull_repo}
\title{Local repository tools}
\arguments{
\item{repos}{GitHub repo names with the form \code{owner/name}.}

\item{repo_dir}{Vector of repo directories or a single directory containing one or more repos.}

\item{message}{commit message}

\item{git}{Path to the local git binary. \code{\link[=require_git]{require_git()}} attempts to
find the git binary based on your `PATH``, it will throw an error if git cannot be found.}

\item{options}{Additional git binary options (e.g. \code{--all}).}

\item{verbose}{Display verbose output.}
}
\description{
The functions provide tools for working with local git repositories, ghclass includes support for following git commands:
\itemize{
\item git clone = \code{clone_repo}
\item git add = \code{add_repo}
\item git commit = \code{commit_repo}
\item git push = \code{push_repo}
\item git pull = \code{pull_repo}
}
}
\examples{
\dontrun{
g = org_repos("Sta323-Sp18","hw3-")
clone_repo(g, "hw3")

pull_repo(g, "hw3")

# After Modifying hw3.Rmd
add_repo("hw3", "hw3.Rmd")
commit_repo("hw3", "Revised homework")
push_repo("hw3")
}

}
