% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_repo_rename.R
\name{local_repo_rename}
\alias{local_repo_rename}
\title{Rename local directories using a vector of patterns and replacements.}
\usage{
local_repo_rename(repo_dir, pattern, replacement)
}
\arguments{
\item{repo_dir}{Character. Vector of repo directories or a single directory containing one or more repos.}

\item{pattern}{Character. One or more regexp patterns to match to directory names.}

\item{replacement}{Character.  One or more text strings containing the replacement value for matched patterns.}
}
\description{
This function is meant to help with renaming local student repos to include
something more useful like \verb{Last, First name} or a unique identifier
for the purposes of ordering repository folders.
}
