% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{mirror_repo}
\alias{mirror_repo}
\title{Mirror repository}
\usage{
mirror_repo(source_repo, target_repo, verbose = TRUE)
}
\arguments{
\item{source_repo}{Character. Address of repository in "owner/name" format.}

\item{target_repo}{Character or vector of characters.  Address of repository in "owner/name" format.}

\item{verbose}{Logical. Display verbose output.}
}
\description{
\code{mirror_repo} mirrors the content of one repository to another repository, or set of repositories. Use the \code{get_repos} function as a wrapper for the target_repo parameter to enable mirroring to multiple repositories.
}
\examples{
\dontrun{
mirror_repo("Sta523/hw1_base", get_repos("Sta523-Fa17","hw1-"))
}

}
