% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R, R/org_exists.R, R/org_repo_stats.R,
%   R/org_repos.R, R/org_team_details.R, R/org_teams.R
\name{org_details}
\alias{org_details}
\alias{org_exists}
\alias{org_repo_stats}
\alias{org_repos}
\alias{org_team_details}
\alias{org_teams}
\title{Obtain details on an organization's repos and teams}
\usage{
org_exists(org)

org_repo_stats(
  org,
  filter = "",
  filter_type = "in:name",
  inc_commits = TRUE,
  inc_issues = TRUE,
  inc_prs = TRUE
)

org_repos(org, filter = NULL, exclude = FALSE, full_repo = TRUE)

org_team_details(org)

org_teams(org, filter = NULL, exclude = FALSE, team_type = c("name", "slug"))
}
\arguments{
\item{org}{Character. Name of the GitHub organization(s).}

\item{filter}{Character. Regular expression pattern for matching (or excluding) results}

\item{filter_type}{Character. One or more GitHub search \verb{in} qualifiers.
See \href{https://help.github.com/en/articles/searching-for-repositories}{documentation}
for more details.}

\item{inc_commits}{Logical. Include commit statistics}

\item{inc_issues}{Logical. Include issue statistics}

\item{inc_prs}{Logical. Include pull request statistics}

\item{exclude}{Logical. Should entries matching the regular expression be excluded or included.}

\item{full_repo}{Logical. Should the full repository address be returned (e.g. \code{owner/repo} instead of just \code{repo})?}

\item{team_type}{Character. Either "slug" if the team names are slugs or "name" if full team names are provided.}
}
\description{
\itemize{
\item \code{org_exists} - returns \code{TRUE} if the organization(s) exist on GitHub and \code{FALSE} otherwise.
\item \code{org_teams} - returns a (filtered) vector of organization teams.
\item \code{org_team_details} - returns a data frame of all organization teams containing identification and permission details.
\item \code{org_repos} - returns a (filtered) vector of organization repositories.
\item \code{org_repo_stats} - returns a tibble of repositories belonging to a GitHub organization along with some
basic statistics about those repositories.
}
}
\examples{
\dontrun{
# Org repos and teams
org_repos("ghclass-test")

org_repos("ghclass-test", filter = "hw1-")

org_teams("ghclass-test")

org_team_details("ghclass-test")
}

}
