% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_members.R
\name{org_members}
\alias{org_members}
\alias{get_member}
\alias{get_members}
\title{Get organization member}
\usage{
org_members(org, filter = NULL, exclude = FALSE,
  include_admins = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression be excluded or included.}

\item{include_admins}{Logical. Should admin users be included in the results.}
}
\description{
\code{org_members} returns a (filtered) vector of organization memebers.
}
\examples{
\dontrun{
org_members("ghclass")
}

}
