% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_repo_stats.R
\name{org_repo_stats}
\alias{org_repo_stats}
\title{Get organization repository}
\usage{
org_repo_stats(org, filter = NULL, exclude = FALSE,
  inc_commits = TRUE, inc_issues = TRUE, inc_prs = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression in \code{filter} be excluded or included?}

\item{inc_commits}{Logical. Include commit statistics}

\item{inc_issues}{Logical. Include issue statistics}

\item{inc_prs}{Logical. Include pull request statistics}
}
\description{
\code{org_repo_stats} returns a tibble of repositories belonging to a GitHub organization along with some
basic statistics about those repositories.
}
\examples{
\dontrun{
org_repo_stats("ghclass-demo")
}

}
