% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_repos.R
\name{org_repos}
\alias{org_repos}
\title{Get organization repository}
\usage{
org_repos(org, filter = NULL, exclude = FALSE, full_repo = TRUE)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{filter}{Character. Regular expression pattern for matching (or excluding) repositories.}

\item{exclude}{Logical. Should entries matching the regular expression in \code{filter} be excluded or included?}

\item{full_repo}{Logical. Should the full repository address be returned (e.g. \code{owner/name} instead of just \code{repo})?}
}
\description{
\code{org_repos} returns a (filtered) vector of repositories belonging to a GitHub organization.
}
\examples{
\dontrun{
org_repos("ghclass")
org_repos("ghclass", "hw1-")
}

}
