% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_team_ids.R
\name{org_team_ids}
\alias{org_team_ids}
\title{Get organization teams and their ids}
\usage{
org_team_ids(org, filter = NULL, exclude = FALSE)
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{filter}{character, regex pattern for matching (or excluding) repos.}

\item{exclude}{logical, should entries matching the regex be excluded or included.}
}
\description{
\code{org_team_ids} returns a (filtered) data frame of teams in the organization with their
names (\code{name}) and unique ids (\code{id}).
}
