% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_assign}
\alias{peer_assign}
\title{Assign file to reviewers}
\usage{
peer_assign(org, roster, file, prefix = "", suffix = "",
  message = "Assigning review", branch = "master", overwrite = FALSE)
}
\arguments{
\item{org}{Character. Name of GitHub Organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{r*} columns that specify review assignments as values of the vector \code{user_random}. See \code{peer_create_feedback}.}

\item{file}{Character. File name or vector of file names to be included.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{message}{Character. Commit message, defaults to "Assigning review."}

\item{branch}{Character. Name of branch the file should be committed to, defaults to \code{master}.}

\item{overwrite}{Logical. Whether existing files in reviewers' repositories should be overwritten, defaults to \code{FALSE}.}
}
\description{
\code{peer_assign} adds files from authors' repositories to reviewers' repositories.
}
