% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_assign}
\alias{peer_assign}
\title{Assign file to reviewers}
\usage{
peer_assign(repo1, repo2, file, folder, message, branch = "master",
  overwrite = F)
}
\arguments{
\item{repo1}{Character. Address of author repository in "owner/name" format.}

\item{repo2}{Character or character vector. Address(es) of author repository or repositories in "owner/name" format.}

\item{file}{Character. File name of file in \code{repo1} to be assigned to \code{repo2}.}

\item{folder}{Character. Folder name of folder to be created in \code{repo2}.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch the file should be committed to, defaults to \code{master}.}
}
\description{
\code{peer_assign} adds files from authors' repositories to reviewers' repositories. The function appends the prefix "author_" to each file. It also removes information from the \code{author} field in RMarkdown files. To ensure anonymity of the author and subsequent workflow, students should be instructed not to change file names.
}
