% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_create_rform}
\alias{peer_create_rform}
\title{Create reviewer feedback form}
\usage{
peer_create_rform(n, title = "Reviewer feedback form",
  file = "rfeedback_blank", output = "github_document",
  write_rmd = TRUE)
}
\arguments{
\item{n}{Numerical. Number of score fields to be included in .Rmd YAML.}

\item{title}{Character. Title of form, defaults to "Reviewer feedback form."}

\item{file}{Character. File name of RMarkdown document, defaults to \code{rfeedback_blank}.}

\item{output}{Character. Output parameter for \code{.Rmd} file, defaults to \code{github_document}.}

\item{write_rmd}{Logical. Whether the feedback form should be saved to a \code{.Rmd} file in the current working directory, defaults to TRUE.}
}
\description{
\code{peer_create_rform} creates blank feedback forms for reviewers based on the user-specified number of questions.
}
