% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{peer_init}
\alias{peer_init}
\title{Initiate peer review repositories}
\usage{
peer_init(org, roster, prefix = "", suffix = "")
}
\arguments{
\item{org}{Character. Name of GitHub Organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \code{rev*} columns that specify review assignments as values of the vector \code{user_random}.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}
}
\description{
\code{peer_init()} initiates peer review repositories. It creates a review repository for each user, adds users to their respective repositories, and applies peer review labels to all repositories (i.e. assignment and review repositories).
}
\examples{
\dontrun{
peer_init(org = "ghclass-test", roster = "roster_test", prefix = "hw2-")
}

}
\seealso{
Other peer review functions: \code{\link{peer_assign}},
  \code{\link{peer_file_add_aut}},
  \code{\link{peer_file_add_rev}},
  \code{\link{peer_form_create_rating}},
  \code{\link{peer_form_create_review}},
  \code{\link{peer_return}},
  \code{\link{peer_roster_create}},
  \code{\link{peer_score_rating}},
  \code{\link{peer_score_review}},
  \code{\link{repo_file_mirror}}
}
\concept{peer review functions}
