% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_get_readme.R
\name{repo_get_readme}
\alias{repo_get_readme}
\title{Low level function for retrieving the README of a GitHub Repository}
\usage{
repo_get_readme(repo, branch = "master")
}
\arguments{
\item{repo}{Character. Address of repository in \code{owner/name} format.}

\item{branch}{Character. Name of branch to use, defaults to "master".}
}
\description{
Low level function for retrieving the README of a GitHub Repository
}
