% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_mirror_template.R
\name{repo_mirror_template}
\alias{repo_mirror_template}
\title{Mirror template repository}
\usage{
repo_mirror_template(template_repo, repo, private = TRUE)
}
\arguments{
\item{template_repo}{Character. Address of template repository in \code{owner/name} format.}

\item{private}{Logical. Should the new repository be private or public.}

\item{new_repo}{Character. One or more repository addresses in \code{owner/name} format.}
}
\description{
\code{repo_mirror_template} mirrors the content of one repository to another repository, or set of
repositories.
}
