% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_style.R
\name{repo_style}
\alias{repo_style}
\title{Style repository with styler}
\usage{
repo_style(repo, files = c("*.R", "*.Rmd"), branch = "styler",
  base = "master", draft = TRUE, create_pull_request = TRUE,
  tag_collaborators = TRUE)
}
\arguments{
\item{repo}{Character. Address of repository in "owner/name" format.}

\item{files}{Character or vector of characters. Names of .R and/or .Rmd files that styler should be applied to.}

\item{branch}{Character. Name of new branch to be created. Default is "styler".}

\item{base}{Character. Name of branch that contains the .R and/or .Rmd files to be reformatted.}

\item{draft}{Logical. Should the pull request be created as a draft pull request? (Draft PRs cannot be merged
until allowed by the author)}

\item{create_pull_request}{Logical. If TRUE, a pull request is created from branch to base.}

\item{tag_collaborators}{Logical. If TRUE, a message with the repository collaborators is displayed.}

\item{git}{Chacacter. Path to the git binary.}
}
\description{
\code{repo_style} implements "non-invasive pretty-printing of R source code" of .R or .Rmd files within a repository using the \code{styler} package and adhering to \code{tidyverse} formatting guidelines.
}
\examples{
\dontrun{
style_repo("Sta523-Fa17/base_hw1", files = c("hw1_sample.Rmd"))
}

}
