% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_unwatch.R
\name{repo_unwatch}
\alias{repo_unwatch}
\alias{unwatch_repo}
\title{Unwatch repository}
\usage{
repo_unwatch(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
Unwatches / unsubscribes from the provided GitHub repositories.
}
\examples{
\dontrun{
repo_unwatch()
repo_unwatch("Sta523-Fa15/hw1-Tim")
}

}
