% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R, R/team_create.R, R/team_delete.R,
%   R/team_invite.R, R/team_members.R, R/team_pending.R, R/team_remove.R,
%   R/team_rename.R, R/team_repos.R
\name{team}
\alias{team}
\alias{team_create}
\alias{team_delete}
\alias{team_invite}
\alias{team_members}
\alias{team_pending}
\alias{team_remove}
\alias{team_rename}
\alias{team_repos}
\title{GitHub team related tools}
\usage{
team_create(
  org,
  team,
  prefix = "",
  suffix = "",
  privacy = c("secret", "closed")
)

team_delete(org, team, team_type = c("name", "slug"), prompt = TRUE)

team_invite(org, user, team, team_type = c("name", "slug"))

team_members(org, team = org_teams(org), team_type = c("name", "slug"))

team_pending(org, team = org_teams(org), team_type = c("name", "slug"))

team_remove(org, user, team, team_type = c("name", "slug"))

team_rename(org, team, new_team, team_type = c("name", "slug"))

team_repos(org, team = org_teams(org), team_type = c("name", "slug"))
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{team}{Character. Name of teams.}

\item{prefix}{Character. Shared prefix.}

\item{suffix}{Character. Shared suffix.}

\item{privacy}{Character. Level of privacy for team, "closed" (visible to all
members of the organization) or "secret" (only visible to organization owners
and members of a team), default is "closed"}

\item{team_type}{Character. Either "slug" if the team names are slugs or "name" if full team names are provided.}

\item{prompt}{Logical. Should the user be prompted before deleting team. Default \code{true}.}

\item{user}{Character. One or more GitHub usernames to invite.}

\item{new_team}{character, new team name.}
}
\description{
\itemize{
\item \code{team_create} - create teams in a GitHub organization
\item \code{team_delete} - delete a team from a GitHub organization.
\item \code{team_rename} - renames an existing team.
\item \code{team_invite} - add members to GitHub organization team(s).
\item \code{team_members} - returns a data frame of team members.
\item \code{team_pending} - returns a data frame of pending team members.
\item \code{team_repos} - returns a data frame of teams and their repos.
}
}
\examples{
\dontrun{
team_delete("ghclass-test", org_teams("ghclass-test", "hw1-"), prompt = FALSE)
team_create("ghclass-test",c("hw1-team01","hw1-team02"))
team_rename("ghclass-test", "hw1-team02", "hw1-team03")
org_teams("ghclass-test", "hw1-")


team_invite("ghclass-test", user = "rundel", team = c("hw1-team01", "hw1-team03", "missing_team"))
team_remove("ghclass-test", user = "rundel", team = "hw1-team01")

team_members("ghclass-test", org_teams("ghclass-test", "hw1-"))
team_pending("ghclass-test", org_teams("ghclass-test", "hw1-"))
team_repos("ghclass-test", org_teams("ghclass-test", "hw1-"))
}

}
