% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_create.R
\name{team_create}
\alias{team_create}
\title{Create team(s)}
\usage{
team_create(org, team, privacy = c("closed", "secret"))
}
\arguments{
\item{org}{character, name of the GitHub organization}

\item{team}{character, listing one or more teams}

\item{privacy}{character, level of privacy of teams, closed (visible to all
members of the organization) or secret (only visible to organization owners
and members of a team), default is closed}
}
\description{
\code{team_create} creates teams in your GitHub organization
}
\examples{
\dontrun{
team_create("ghclass",c("team01","team01"))
}

}
\seealso{
Other github organization team related functions: \code{\link{get_teams}},
  \code{\link{team_delete}}, \code{\link{team_invite}},
  \code{\link{team_members}},
  \code{\link{team_pending_members}},
  \code{\link{team_rename}}
}
\concept{github organization team related functions}
