% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_members.R
\name{team_members}
\alias{team_members}
\title{Get team members}
\usage{
team_members(org, team = org_teams(org))
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{team}{character or data frame, listing one or more team}
}
\description{
\code{team_members} returns a data frame of teams and their members.
}
\examples{
\dontrun{
team_members("ghclass-test",c("team01","team02"))
}

}
