% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_members.R
\name{team_members}
\alias{team_members}
\title{Get team members}
\usage{
team_members(org, team = get_teams(org))
}
\arguments{
\item{org}{character, name of the GitHub organization.}

\item{team}{character or data frame, listing one or more team}
}
\description{
\code{team_members} returns a data frame of teams and their members.
}
\examples{
\dontrun{
team_members("ghclass-test",c("team01","team02"))
}

}
\seealso{
Other github organization team related functions: \code{\link{get_teams}},
  \code{\link{team_create}}, \code{\link{team_delete}},
  \code{\link{team_invite}},
  \code{\link{team_pending_members}},
  \code{\link{team_rename}}
}
\concept{github organization team related functions}
