% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{test_github_token}
\alias{test_github_token}
\title{Test github token}
\usage{
test_github_token(token = get_github_token())
}
\arguments{
\item{token}{github api personal access token}
}
\description{
\code{test_github_token} checks if a token is valid by attempting to authenticate with the GitHub api.
}
\examples{
\dontrun{
test_github_token()
test_github_token("bad_token")
}

}
\seealso{
Other authentication functions: \code{\link{get_github_token}},
  \code{\link{set_github_token}}
}
\concept{authentication functions}
