% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{test_github_token}
\alias{test_github_token}
\title{Test github token}
\usage{
test_github_token(token)
}
\arguments{
\item{token}{character or missing, if missing the token is obtained using \code{get_github_token}.}
}
\description{
\code{test_github_token} checks if a token is valid by attempting to authenticate with the GitHub api.
}
\examples{
\dontrun{
test_github_token()
test_github_token("bad_token")
}

}
\seealso{
Other authentication functions: \code{\link{get_github_token}},
  \code{\link{get_wercker_token}},
  \code{\link{set_github_token}},
  \code{\link{set_wercker_token}},
  \code{\link{test_wercker_token}}
}
