% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{test_wercker_token}
\alias{test_wercker_token}
\title{Test wercker token}
\usage{
test_wercker_token(token)
}
\arguments{
\item{token}{character or missing, if missing the token is obtained using \code{get_wercker_token}.}
}
\description{
\code{test_wercker_token} checks if a token is valid by attempting to authenticate with the Wercker's api.
}
\examples{
\dontrun{
test_wercker_token()
test_wercker_token("bad_token")
}

}
\seealso{
Other authentication functions: \code{\link{get_github_token}},
  \code{\link{get_wercker_token}},
  \code{\link{set_github_token}},
  \code{\link{set_wercker_token}},
  \code{\link{test_github_token}}
}
\concept{authentication functions}
