% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watching.R
\name{unwatch_repo}
\alias{unwatch_repo}
\title{Unwatch repository}
\usage{
unwatch_repo(repo)
}
\arguments{
\item{repo}{github repository address in \code{owner/repo} format}
}
\description{
Unwatches / unsubscribes from the provided GitHub repositories.
}
\examples{
\dontrun{
unwatch_repo()
unwatch_repo("Sta523-Fa15/hw1-Tim")
}

}
\seealso{
Other notification functions: \code{\link{get_watching}},
  \code{\link{ignore_repo}}, \code{\link{watch_repo}}
}
\concept{notification functions}
